Crate objc2_foundation

Bindings to the Foundation framework

See Apple’s docs and the general docs on framework crates for more information.

This is the std equivalent for Objective-C, containing essential data types, collections, and operating-system services.

Rust vs. Objective-C types

A quick overview of some types you will encounter often in Objective-C, and their approximate Rust equivalent.

Objective-C	(approximately) equivalent Rust

NSData*	Rc<[u8]>

NSMutableData*	Rc<Cell<Vec<u8>>>

NSString*	Rc<str>

NSMutableString*	Rc<Cell<String>>

NSValue*	Rc<dyn Any>

NSNumber*	Arc<enum { I8(i8), U8(u8), I16(i16), ... }>

NSError*	Arc<dyn Error + Send + Sync>

NSException*	Arc<dyn Error + Send + Sync>

NSRange	ops::Range<usize>

NSComparisonResult	cmp::Ordering

NSEnumerator<T>*	Rc<dyn Iterator<Item = Retained<T>>>

NSCopying*	Rc<dyn Clone>

NSArray<T>*	Rc<[Retained<T>]>

NSMutableArray<T>*	Rc<Cell<Vec<Retained<T>>>>

NSDictionary<K, V>*	Rc<HashMap<Retained<K>, Retained<V>>>

NSMutableDictionary<K, V>*	Rc<Cell<HashMap<Retained<K>, Retained<V>>>>

Note, in particular, that all “Mutable” variants use interior mutability, and that some things are thread-safe (Arc), while others are not (Rc).

Examples

Basic usage of a few Foundation types.

$ cargo add objc2-foundation

use objc2_foundation::{ns_string, NSArray, NSDictionary, NSObject};

// Create and compare NSObjects
let obj = NSObject::new();
#[allow(clippy::eq_op)]
{
    println!("{obj:?} == {obj:?}? {:?}", obj == obj);
}

let obj2 = NSObject::new();
println!("{obj:?} == {obj2:?}? {:?}", obj == obj2);

// Create an NSArray from a Vec
let objs = vec![obj, obj2];
let array = NSArray::from_retained_slice(&objs);
for obj in array.iter() {
    println!("{obj:?}");
}
println!("{}", array.len());

// Turn the NSArray back into a Vec
let mut objs = array.to_vec();
let obj = objs.pop().unwrap();

// Create a static NSString
let string = ns_string!("Hello, world!");
// And use the `ToString` implementation to convert it into a string
let _s = string.to_string();
// Or use the `Display` implementation directly
println!("{string}");

// Create a dictionary mapping strings to objects
let keys = &[string];
let objects = &[&*obj];
let dict = NSDictionary::from_slices(keys, objects);
println!("{:?}", dict.objectForKey(string));
println!("{}", dict.len());

Modules

array

Utilities for the NSArray and NSMutableArray classes.

dictionary

Utilities for the NSDictionary and NSMutableDictionary classes.

enumerator

Utilities for the NSEnumerator class.

set

Utilities for the NSSet and NSMutableSet classes.

Macros

ns_string

Create a NSString from a static str.

Structs

MainThreadMarker

A marker type for functionality only available on the main thread.

NSAlignmentOptions

Apple’s documentation

NSArray

************** Immutable Array ***************

NSBinarySearchingOptions

Apple’s documentation

NSBundle

Apple’s documentation

NSBundleResourceRequest

Apple’s documentation

NSCoder

Apple’s documentation

NSConstantString

Apple’s documentation

NSCountedSet

************** Counted Set ***************

NSData

************** Immutable Data ***************

NSDataBase64DecodingOptions

Apple’s documentation

NSDataBase64EncodingOptions

************** Base 64 Options ***************

NSDataCompressionAlgorithm

Apple’s documentation

NSDataReadingOptions

************** Read/Write Options ***************

NSDataSearchOptions

************** Data Search Options ***************

NSDataWritingOptions

Apple’s documentation

NSDecodingFailurePolicy

Describes the action an NSCoder should take when it encounters decode failures (e.g. corrupt data) for non-TopLevel decodes.

NSDictionary

************** Immutable Dictionary ***************

NSEdgeInsets

Apple’s documentation

NSEnumerator

Apple’s documentation

NSError

Apple’s documentation

NSFastEnumerationState

NSFileSecurity

Apple’s documentation

NSFileWrapper

Apple’s documentation

NSFileWrapperReadingOptions

Apple’s documentation

NSFileWrapperWritingOptions

Apple’s documentation

NSItemProvider

Apple’s documentation

NSItemProviderErrorCode

Apple’s documentation

NSItemProviderFileOptions

Apple’s documentation

NSItemProviderRepresentationVisibility

Apple’s documentation

NSLocale

Apple’s documentation

NSLocaleLanguageDirection

Apple’s documentation

NSMutableArray

************** Mutable Array ***************

NSMutableData

************** Mutable Data ***************

NSMutableDictionary

************** Mutable Dictionary ***************

NSMutableSet

************** Mutable Set ***************

NSMutableString

Apple’s documentation

NSNumber

Apple’s documentation

NSObject

The root class of most Objective-C class hierarchies.

NSPurgeableData

************** Purgeable Data ***************

NSRectEdge

NSSet

************** Immutable Set ***************

NSSimpleCString

Apple’s documentation

NSString

Apple’s documentation

NSStringCompareOptions

Apple’s documentation

NSStringEncodingConversionOptions

Apple’s documentation

NSStringEnumerationOptions

Apple’s documentation

NSURL

Apple’s documentation

NSURLBookmarkCreationOptions

Apple’s documentation

NSURLBookmarkResolutionOptions

Apple’s documentation

NSURLComponents

Apple’s documentation

NSURLQueryItem

Apple’s documentation

NSValue

Apple’s documentation

Constants

NSASCIIStringEncoding

Apple’s documentation

NSBundleExecutableArchitectureARM64

Apple’s documentation

NSBundleExecutableArchitectureI386

Apple’s documentation

NSBundleExecutableArchitecturePPC

Apple’s documentation

NSBundleExecutableArchitecturePPC64

Apple’s documentation

NSBundleExecutableArchitectureX86_64

Apple’s documentation

NSISO2022JPStringEncoding

Apple’s documentation

NSISOLatin1StringEncoding

Apple’s documentation

NSISOLatin2StringEncoding

Apple’s documentation

NSJapaneseEUCStringEncoding

Apple’s documentation

NSMacOSRomanStringEncoding

Apple’s documentation

NSNEXTSTEPStringEncoding

Apple’s documentation

NSNonLossyASCIIStringEncoding

Apple’s documentation

NSShiftJISStringEncoding

Apple’s documentation

NSSymbolStringEncoding

Apple’s documentation

NSUTF8StringEncoding

Apple’s documentation

NSUTF16BigEndianStringEncoding

Apple’s documentation

NSUTF16LittleEndianStringEncoding

Apple’s documentation

NSUTF16StringEncoding

Apple’s documentation

NSUTF32BigEndianStringEncoding

Apple’s documentation

NSUTF32LittleEndianStringEncoding

Apple’s documentation

NSUTF32StringEncoding

Apple’s documentation

NSUnicodeStringEncoding

Apple’s documentation

NSWindowsCP1250StringEncoding

Apple’s documentation

NSWindowsCP1251StringEncoding

Apple’s documentation

NSWindowsCP1252StringEncoding

Apple’s documentation

NSWindowsCP1253StringEncoding

Apple’s documentation

NSWindowsCP1254StringEncoding

Apple’s documentation

Statics

NSBuddhistCalendar⚠Deprecated

Apple’s documentation

NSBundleResourceRequestLoadingPriorityUrgent⚠

Apple’s documentation

NSChineseCalendar⚠Deprecated

Apple’s documentation

NSCocoaErrorDomain⚠

Apple’s documentation

NSDebugDescriptionErrorKey⚠

Apple’s documentation

NSEdgeInsetsZero⚠

Apple’s documentation

NSExtensionJavaScriptFinalizeArgumentKey⚠

Apple’s documentation

NSExtensionJavaScriptPreprocessingResultsKey⚠

Apple’s documentation

NSFilePathErrorKey⚠

Apple’s documentation

NSGregorianCalendar⚠Deprecated

Apple’s documentation

NSHebrewCalendar⚠Deprecated

Apple’s documentation

NSHelpAnchorErrorKey⚠

Apple’s documentation

NSISO8601Calendar⚠Deprecated

Apple’s documentation

NSIndianCalendar⚠Deprecated

Apple’s documentation

NSIslamicCalendar⚠Deprecated

Apple’s documentation

NSIslamicCivilCalendar⚠Deprecated

Apple’s documentation

NSItemProviderErrorDomain⚠

Apple’s documentation

NSItemProviderPreferredImageSizeKey⚠

Apple’s documentation

NSJapaneseCalendar⚠Deprecated

Apple’s documentation

NSLoadedClasses⚠

Apple’s documentation

NSLocaleAlternateQuotationBeginDelimiterKey⚠

Apple’s documentation

NSLocaleAlternateQuotationEndDelimiterKey⚠

Apple’s documentation

NSLocaleCalendar⚠

Apple’s documentation

NSLocaleCollationIdentifier⚠

Apple’s documentation

NSLocaleCollatorIdentifier⚠

Apple’s documentation

NSLocaleCountryCode⚠

Apple’s documentation

NSLocaleCurrencyCode⚠

Apple’s documentation

NSLocaleCurrencySymbol⚠

Apple’s documentation

NSLocaleDecimalSeparator⚠

Apple’s documentation

NSLocaleExemplarCharacterSet⚠

Apple’s documentation

NSLocaleGroupingSeparator⚠

Apple’s documentation

NSLocaleIdentifier⚠

Apple’s documentation

NSLocaleLanguageCode⚠

Apple’s documentation

NSLocaleMeasurementSystem⚠

Apple’s documentation

NSLocaleQuotationBeginDelimiterKey⚠

Apple’s documentation

NSLocaleQuotationEndDelimiterKey⚠

Apple’s documentation

NSLocaleScriptCode⚠

Apple’s documentation

NSLocaleUsesMetricSystem⚠

Apple’s documentation

NSLocaleVariantCode⚠

Apple’s documentation

NSLocalizedDescriptionKey⚠

Apple’s documentation

NSLocalizedFailureErrorKey⚠

Apple’s documentation

NSLocalizedFailureReasonErrorKey⚠

Apple’s documentation

NSLocalizedRecoveryOptionsErrorKey⚠

Apple’s documentation

NSLocalizedRecoverySuggestionErrorKey⚠

Apple’s documentation

NSMachErrorDomain⚠

Apple’s documentation

NSMultipleUnderlyingErrorsKey⚠

Apple’s documentation

NSOSStatusErrorDomain⚠

Apple’s documentation

NSPOSIXErrorDomain⚠

Apple’s documentation

NSPersianCalendar⚠Deprecated

Apple’s documentation

NSRecoveryAttempterErrorKey⚠

Apple’s documentation

NSRepublicOfChinaCalendar⚠Deprecated

Apple’s documentation

NSStringEncodingDetectionAllowLossyKey⚠

Apple’s documentation

NSStringEncodingDetectionDisallowedEncodingsKey⚠

Apple’s documentation

NSStringEncodingDetectionFromWindowsKey⚠

Apple’s documentation

NSStringEncodingDetectionLikelyLanguageKey⚠

Apple’s documentation

NSStringEncodingDetectionLossySubstitutionKey⚠

Apple’s documentation

NSStringEncodingDetectionSuggestedEncodingsKey⚠

Apple’s documentation

NSStringEncodingDetectionUseOnlySuggestedEncodingsKey⚠

Apple’s documentation

NSStringEncodingErrorKey⚠

Apple’s documentation

NSStringTransformFullwidthToHalfwidth⚠

Apple’s documentation

NSStringTransformHiraganaToKatakana⚠

Apple’s documentation

NSStringTransformLatinToArabic⚠

Apple’s documentation

NSStringTransformLatinToCyrillic⚠

Apple’s documentation

NSStringTransformLatinToGreek⚠

Apple’s documentation

NSStringTransformLatinToHangul⚠

Apple’s documentation

NSStringTransformLatinToHebrew⚠

Apple’s documentation

NSStringTransformLatinToHiragana⚠

Apple’s documentation

NSStringTransformLatinToKatakana⚠

Apple’s documentation

NSStringTransformLatinToThai⚠

Apple’s documentation

NSStringTransformMandarinToLatin⚠

Apple’s documentation

NSStringTransformStripCombiningMarks⚠

Apple’s documentation

NSStringTransformStripDiacritics⚠

Apple’s documentation

NSStringTransformToLatin⚠

Apple’s documentation

NSStringTransformToUnicodeName⚠

Apple’s documentation

NSStringTransformToXMLHex⚠

Apple’s documentation

NSThumbnail1024x1024SizeKey⚠Deprecated

Apple’s documentation

NSURLAddedToDirectoryDateKey⚠

Apple’s documentation

NSURLApplicationIsScriptableKey⚠

Apple’s documentation

NSURLAttributeModificationDateKey⚠

Apple’s documentation

NSURLCanonicalPathKey⚠

Apple’s documentation

NSURLContentAccessDateKey⚠

Apple’s documentation

NSURLContentModificationDateKey⚠

Apple’s documentation

NSURLContentTypeKey⚠

Apple’s documentation

NSURLCreationDateKey⚠

Apple’s documentation

NSURLCustomIconKey⚠

Apple’s documentation

NSURLDirectoryEntryCountKey⚠

Apple’s documentation

NSURLDocumentIdentifierKey⚠

Apple’s documentation

NSURLEffectiveIconKey⚠

Apple’s documentation

NSURLErrorKey⚠

Apple’s documentation

NSURLFileAllocatedSizeKey⚠

Apple’s documentation

NSURLFileContentIdentifierKey⚠

Apple’s documentation

NSURLFileIdentifierKey⚠

Apple’s documentation

NSURLFileProtectionComplete⚠

Apple’s documentation

NSURLFileProtectionCompleteUnlessOpen⚠

Apple’s documentation

NSURLFileProtectionCompleteUntilFirstUserAuthentication⚠

Apple’s documentation

NSURLFileProtectionCompleteWhenUserInactive⚠

Apple’s documentation

NSURLFileProtectionKey⚠

Apple’s documentation

NSURLFileProtectionNone⚠

Apple’s documentation

NSURLFileResourceIdentifierKey⚠

Apple’s documentation

NSURLFileResourceTypeBlockSpecial⚠

Apple’s documentation

NSURLFileResourceTypeCharacterSpecial⚠

Apple’s documentation

NSURLFileResourceTypeDirectory⚠

Apple’s documentation

NSURLFileResourceTypeKey⚠

Apple’s documentation

NSURLFileResourceTypeNamedPipe⚠

Apple’s documentation

NSURLFileResourceTypeRegular⚠

Apple’s documentation

NSURLFileResourceTypeSocket⚠

Apple’s documentation

NSURLFileResourceTypeSymbolicLink⚠

Apple’s documentation

NSURLFileResourceTypeUnknown⚠

Apple’s documentation

NSURLFileScheme⚠

Apple’s documentation

NSURLFileSecurityKey⚠

Apple’s documentation

NSURLFileSizeKey⚠

Apple’s documentation

NSURLGenerationIdentifierKey⚠

Apple’s documentation

NSURLHasHiddenExtensionKey⚠

Apple’s documentation

NSURLIsAliasFileKey⚠

Apple’s documentation

NSURLIsApplicationKey⚠

Apple’s documentation

NSURLIsDirectoryKey⚠

Apple’s documentation

NSURLIsExcludedFromBackupKey⚠

Apple’s documentation

NSURLIsExecutableKey⚠

Apple’s documentation

NSURLIsHiddenKey⚠

Apple’s documentation

NSURLIsMountTriggerKey⚠

Apple’s documentation

NSURLIsPackageKey⚠

Apple’s documentation

NSURLIsPurgeableKey⚠

Apple’s documentation

NSURLIsReadableKey⚠

Apple’s documentation

NSURLIsRegularFileKey⚠

Apple’s documentation

NSURLIsSparseKey⚠

Apple’s documentation

NSURLIsSymbolicLinkKey⚠

Apple’s documentation

NSURLIsSystemImmutableKey⚠

Apple’s documentation

NSURLIsUbiquitousItemKey⚠

Apple’s documentation

NSURLIsUserImmutableKey⚠

Apple’s documentation

NSURLIsVolumeKey⚠

Apple’s documentation

NSURLIsWritableKey⚠

Apple’s documentation

NSURLKeysOfUnsetValuesKey⚠

Apple’s documentation

NSURLLabelColorKey⚠

Apple’s documentation

NSURLLabelNumberKey⚠

Apple’s documentation

NSURLLinkCountKey⚠

Apple’s documentation

NSURLLocalizedLabelKey⚠

Apple’s documentation

NSURLLocalizedNameKey⚠

Apple’s documentation

NSURLLocalizedTypeDescriptionKey⚠

Apple’s documentation

NSURLMayHaveExtendedAttributesKey⚠

Apple’s documentation

NSURLMayShareFileContentKey⚠

Apple’s documentation

NSURLNameKey⚠

Apple’s documentation

NSURLParentDirectoryURLKey⚠

Apple’s documentation

NSURLPathKey⚠

Apple’s documentation

NSURLPreferredIOBlockSizeKey⚠

Apple’s documentation

NSURLQuarantinePropertiesKey⚠

Apple’s documentation

NSURLTagNamesKey⚠

Apple’s documentation

NSURLThumbnailDictionaryKey⚠Deprecated

Apple’s documentation

NSURLThumbnailKey⚠Deprecated

Apple’s documentation

NSURLTotalFileAllocatedSizeKey⚠

Apple’s documentation

NSURLTotalFileSizeKey⚠

Apple’s documentation

NSURLTypeIdentifierKey⚠Deprecated

Apple’s documentation

NSURLUbiquitousItemContainerDisplayNameKey⚠

Apple’s documentation

NSURLUbiquitousItemDownloadRequestedKey⚠

Apple’s documentation

NSURLUbiquitousItemDownloadingErrorKey⚠

Apple’s documentation

NSURLUbiquitousItemDownloadingStatusCurrent⚠

Apple’s documentation

NSURLUbiquitousItemDownloadingStatusDownloaded⚠

Apple’s documentation

NSURLUbiquitousItemDownloadingStatusKey⚠

Apple’s documentation

NSURLUbiquitousItemDownloadingStatusNotDownloaded⚠

Apple’s documentation

NSURLUbiquitousItemHasUnresolvedConflictsKey⚠

Apple’s documentation

NSURLUbiquitousItemIsDownloadedKey⚠Deprecated

Apple’s documentation

NSURLUbiquitousItemIsDownloadingKey⚠

Apple’s documentation

NSURLUbiquitousItemIsExcludedFromSyncKey⚠

Apple’s documentation

NSURLUbiquitousItemIsSharedKey⚠

Apple’s documentation

NSURLUbiquitousItemIsSyncPausedKey⚠

Apple’s documentation

NSURLUbiquitousItemIsUploadedKey⚠

Apple’s documentation

NSURLUbiquitousItemIsUploadingKey⚠

Apple’s documentation

NSURLUbiquitousItemPercentDownloadedKey⚠Deprecated

Apple’s documentation

NSURLUbiquitousItemPercentUploadedKey⚠Deprecated

Apple’s documentation

NSURLUbiquitousItemSupportedSyncControlsKey⚠

Apple’s documentation

NSURLUbiquitousItemUploadingErrorKey⚠

Apple’s documentation

NSURLUbiquitousSharedItemCurrentUserPermissionsKey⚠

Apple’s documentation

NSURLUbiquitousSharedItemCurrentUserRoleKey⚠

Apple’s documentation

NSURLUbiquitousSharedItemMostRecentEditorNameComponentsKey⚠

Apple’s documentation

NSURLUbiquitousSharedItemOwnerNameComponentsKey⚠

Apple’s documentation

NSURLUbiquitousSharedItemPermissionsReadOnly⚠

Apple’s documentation

NSURLUbiquitousSharedItemPermissionsReadWrite⚠

Apple’s documentation

NSURLUbiquitousSharedItemRoleOwner⚠

Apple’s documentation

NSURLUbiquitousSharedItemRoleParticipant⚠

Apple’s documentation

NSURLVolumeAvailableCapacityForImportantUsageKey⚠

Apple’s documentation

NSURLVolumeAvailableCapacityForOpportunisticUsageKey⚠

Apple’s documentation

NSURLVolumeAvailableCapacityKey⚠

Apple’s documentation

NSURLVolumeCreationDateKey⚠

Apple’s documentation

NSURLVolumeIdentifierKey⚠

Apple’s documentation

NSURLVolumeIsAutomountedKey⚠

Apple’s documentation

NSURLVolumeIsBrowsableKey⚠

Apple’s documentation

NSURLVolumeIsEjectableKey⚠

Apple’s documentation

NSURLVolumeIsEncryptedKey⚠

Apple’s documentation

NSURLVolumeIsInternalKey⚠

Apple’s documentation

NSURLVolumeIsJournalingKey⚠

Apple’s documentation

NSURLVolumeIsLocalKey⚠

Apple’s documentation

NSURLVolumeIsReadOnlyKey⚠

Apple’s documentation

NSURLVolumeIsRemovableKey⚠

Apple’s documentation

NSURLVolumeIsRootFileSystemKey⚠

Apple’s documentation

NSURLVolumeLocalizedFormatDescriptionKey⚠

Apple’s documentation

NSURLVolumeLocalizedNameKey⚠

Apple’s documentation

NSURLVolumeMaximumFileSizeKey⚠

Apple’s documentation

NSURLVolumeMountFromLocationKey⚠

Apple’s documentation

NSURLVolumeNameKey⚠

Apple’s documentation

NSURLVolumeResourceCountKey⚠

Apple’s documentation

NSURLVolumeSubtypeKey⚠

Apple’s documentation

NSURLVolumeSupportsAccessPermissionsKey⚠

Apple’s documentation

NSURLVolumeSupportsAdvisoryFileLockingKey⚠

Apple’s documentation

NSURLVolumeSupportsCasePreservedNamesKey⚠

Apple’s documentation

NSURLVolumeSupportsCaseSensitiveNamesKey⚠

Apple’s documentation

NSURLVolumeSupportsCompressionKey⚠

Apple’s documentation

NSURLVolumeSupportsExclusiveRenamingKey⚠

Apple’s documentation

NSURLVolumeSupportsExtendedSecurityKey⚠

Apple’s documentation

NSURLVolumeSupportsFileCloningKey⚠

Apple’s documentation

NSURLVolumeSupportsFileProtectionKey⚠

Apple’s documentation

NSURLVolumeSupportsHardLinksKey⚠

Apple’s documentation

NSURLVolumeSupportsImmutableFilesKey⚠

Apple’s documentation

NSURLVolumeSupportsJournalingKey⚠

Apple’s documentation

NSURLVolumeSupportsPersistentIDsKey⚠

Apple’s documentation

NSURLVolumeSupportsRenamingKey⚠

Apple’s documentation

NSURLVolumeSupportsRootDirectoryDatesKey⚠

Apple’s documentation

NSURLVolumeSupportsSparseFilesKey⚠

Apple’s documentation

NSURLVolumeSupportsSwapRenamingKey⚠

Apple’s documentation

NSURLVolumeSupportsSymbolicLinksKey⚠

Apple’s documentation

NSURLVolumeSupportsVolumeSizesKey⚠

Apple’s documentation

NSURLVolumeSupportsZeroRunsKey⚠

Apple’s documentation

NSURLVolumeTotalCapacityKey⚠

Apple’s documentation

NSURLVolumeTypeNameKey⚠

Apple’s documentation

NSURLVolumeURLForRemountingKey⚠

Apple’s documentation

NSURLVolumeURLKey⚠

Apple’s documentation

NSURLVolumeUUIDStringKey⚠

Apple’s documentation

NSUnderlyingErrorKey⚠

Apple’s documentation

NSZeroPoint⚠

Apple’s documentation

NSZeroRect⚠

Apple’s documentation

NSZeroSize⚠

Apple’s documentation

Traits

CopyingHelper

A helper type for implementing NSCopying.

MutableCopyingHelper

A helper type for implementing NSMutableCopying.

NSCoding

Apple’s documentation

NSCopying

A protocol to provide functional copies of objects.

NSDiscardableContent

********* Discardable Content **********

NSFastEnumeration

Apple’s documentation

NSItemProviderReading

Apple’s documentation

NSItemProviderWriting

Apple’s documentation

NSMutableCopying

A protocol to provide mutable copies of objects.

NSObjectNSCoderMethods

Category “NSCoderMethods” on NSObject.

NSObjectNSDiscardableContentProxy

Category “NSDiscardableContentProxy” on NSObject.

NSObjectNSErrorRecoveryAttempting

Category “NSErrorRecoveryAttempting” on NSObject.

NSObjectProtocol

The methods that are fundamental to most Objective-C objects.

NSSecureCoding

Apple’s documentation

Functions

NSContainsRect

NSDeallocateObject⚠

Safety

NSDecrementExtraRefCountWasZero⚠

Safety

NSDivideRect⚠

Safety

NSEdgeInsetsEqualDeprecated

NSEqualPoints

NSEqualRects

NSEqualSizes

NSExtraRefCount⚠

Safety

NSIncrementExtraRefCount⚠

Safety

NSInsetRect

NSIntegralRect

NSIntegralRectWithOptions

NSIntersectionRect

NSIntersectsRect

NSIsEmptyRect

NSMouseInRect

NSOffsetRect

NSPointFromString

NSPointInRect

NSRectFromString

NSSizeFromString

NSStringFromPointDeprecated

NSStringFromRectDeprecated

NSStringFromSizeDeprecated

NSUnionRect

NXReadNSObjectFromCoder⚠Deprecated

Safety

Type Aliases

NSErrorDomain

Apple’s documentation

NSErrorUserInfoKey

Apple’s documentation

NSInteger

A signed integer value type.

NSLocaleKey

Apple’s documentation

NSPoint

A point in a Cartesian coordinate system.

NSPointArray

Apple’s documentation

NSPointPointer

Apple’s documentation

NSRect

A rectangle.

NSRectArray

Apple’s documentation

NSRectPointer

Apple’s documentation

NSSize

A two-dimensional size.

NSSizeArray

Apple’s documentation

NSSizePointer

Apple’s documentation

NSStringEncoding

Apple’s documentation

NSStringEncodingDetectionOptionsKey

Apple’s documentation

NSStringTransform

Apple’s documentation

NSUInteger

Describes an unsigned integer.

NSURLBookmarkFileCreationOptions

Apple’s documentation

NSURLFileProtectionType

Apple’s documentation

NSURLFileResourceType

Apple’s documentation

NSURLResourceKey

Apple’s documentation

NSURLThumbnailDictionaryItem

Apple’s documentation

NSURLUbiquitousItemDownloadingStatus

Apple’s documentation

NSURLUbiquitousSharedItemPermissions

Apple’s documentation

NSURLUbiquitousSharedItemRole

Apple’s documentation

unichar

Apple’s documentation