objc2_foundation

Module dictionary

Utilities for the NSDictionary and NSMutableDictionary classes.

Structs

IntoValues

A consuming iterator over the values of a NSDictionary.

Keys

An iterator over the keys of a NSDictionary.

KeysRetained

An iterator that retains the keys of a NSDictionary.

Values

An iterator over the values of a NSDictionary.

ValuesMut

A mutable iterator over the values of a NSDictionary.

ValuesRetained

A iterator that retains the values of a NSDictionary.