objc2_foundation

Struct NSDictionary

pub struct NSDictionary<KeyType: ?Sized = AnyObject, ObjectType: ?Sized = AnyObject> { /* private fields */ }

Implementations

impl<K: Message + Eq + Hash + HasStableHash, V: Message> NSDictionary<K, V>

pub fn from_vec<Q>(keys: &[&Q], objects: Vec<Retained<V>>) -> Retained<Self>
where Q: Message + NSCopying + CounterpartOrSelf<Immutable = K>,

pub fn from_id_slice<Q>(keys: &[&Q], objects: &[Retained<V>]) -> Retained<Self>
where Q: Message + NSCopying + CounterpartOrSelf<Immutable = K>, V: IsIdCloneable,

pub fn from_slice<Q>(keys: &[&Q], objects: &[&V]) -> Retained<Self>
where Q: Message + NSCopying + CounterpartOrSelf<Immutable = K>, V: IsRetainable,

impl<K: Message + Eq + Hash, V: Message> NSDictionary<K, V>

pub fn get(&self, key: &K) -> Option<&V>

pub fn get_retained(&self, key: &K) -> Option<Retained<V>>
where V: IsIdCloneable,

pub fn get_mut(&mut self, key: &K) -> Option<&mut V>
where V: IsMutable,

Returns a mutable reference to the value corresponding to the key.

Examples

 use objc2_foundation::{ns_string, NSMutableDictionary, NSMutableString, NSString};

 let mut dict = NSMutableDictionary::new();
 dict.insert_id(ns_string!("one"), NSMutableString::new());
 println!("{:?}", dict.get_mut(ns_string!("one")));

impl<K: Message, V: Message> NSDictionary<K, V>

pub fn len(&self) -> usize

pub fn is_empty(&self) -> bool

pub fn keys_vec(&self) -> Vec<&K>

pub fn values_vec(&self) -> Vec<&V>

pub fn values_vec_mut(&mut self) -> Vec<&mut V>
where V: IsMutable,

Returns a vector of mutable references to the values in the dictionary.

Examples

use objc2_foundation::{ns_string, NSMutableDictionary, NSMutableString, NSString};

let mut dict = NSMutableDictionary::new();
dict.insert_id(ns_string!("one"), NSMutableString::from_str("two"));
for val in dict.values_mut() {
    println!("{:?}", val);
}

pub fn to_vecs(&self) -> (Vec<&K>, Vec<&V>)

pub fn to_array(&self) -> Retained<NSArray<V>>
where V: IsIdCloneable,

Returns an NSArray containing the dictionary’s values.

Examples

use objc2_foundation::{ns_string, NSMutableDictionary, NSObject, NSString};

let mut dict = NSMutableDictionary::new();
dict.insert_id(ns_string!("one"), NSObject::new());
let array = dict.to_array();
assert_eq!(array.len(), 1);

impl<K: Message, V: Message> NSDictionary<K, V>

pub fn keys(&self) -> Keys<'_, K, V>

pub fn keys_retained(&self) -> KeysRetained<'_, K, V>
where K: IsIdCloneable,

pub fn values(&self) -> Values<'_, K, V>

pub fn values_mut(&mut self) -> ValuesMut<'_, K, V>
where V: IsMutable,

pub fn values_retained(&self) -> ValuesRetained<'_, K, V>
where V: IsIdCloneable,

pub fn into_values(this: Retained<Self>) -> IntoValues<K, V>
where V: IsIdCloneable,

impl<KeyType: Message, ObjectType: Message> NSDictionary<KeyType, ObjectType>

pub fn count(&self) -> NSUInteger

pub unsafe fn objectForKey( &self, a_key: &KeyType, ) -> Option<Retained<ObjectType>>

pub unsafe fn keyEnumerator(&self) -> Retained<NSEnumerator<KeyType>>

pub fn init(this: Allocated<Self>) -> Retained<Self>

pub unsafe fn initWithObjects_forKeys_count( this: Allocated<Self>, objects: *mut NonNull<ObjectType>, keys: *mut NonNull<ProtocolObject<dyn NSCopying>>, cnt: NSUInteger, ) -> Retained<Self>

pub unsafe fn initWithCoder( this: Allocated<Self>, coder: &NSCoder, ) -> Option<Retained<Self>>

impl<KeyType: Message, ObjectType: Message> NSDictionary<KeyType, ObjectType>

Methods declared on superclass NSObject

pub fn new() -> Retained<Self>

impl<KeyType: Message, ObjectType: Message> NSDictionary<KeyType, ObjectType>

NSExtendedDictionary

pub unsafe fn allKeys(&self) -> Retained<NSArray<KeyType>>

pub unsafe fn allKeysForObject( &self, an_object: &ObjectType, ) -> Retained<NSArray<KeyType>>

pub unsafe fn allValues(&self) -> Retained<NSArray<ObjectType>>

pub unsafe fn description(&self) -> Retained<NSString>

pub unsafe fn descriptionInStringsFileFormat(&self) -> Retained<NSString>

pub unsafe fn descriptionWithLocale( &self, locale: Option<&AnyObject>, ) -> Retained<NSString>

pub unsafe fn descriptionWithLocale_indent( &self, locale: Option<&AnyObject>, level: NSUInteger, ) -> Retained<NSString>

pub unsafe fn isEqualToDictionary( &self, other_dictionary: &NSDictionary<KeyType, ObjectType>, ) -> bool

pub unsafe fn objectEnumerator(&self) -> Retained<NSEnumerator<ObjectType>>

pub unsafe fn objectsForKeys_notFoundMarker( &self, keys: &NSArray<KeyType>, marker: &ObjectType, ) -> Retained<NSArray<ObjectType>>

pub unsafe fn writeToURL_error( &self, url: &NSURL, ) -> Result<(), Retained<NSError>>

pub unsafe fn keysSortedByValueUsingSelector( &self, comparator: Sel, ) -> Retained<NSArray<KeyType>>

pub unsafe fn getObjects_andKeys_count( &self, objects: *mut NonNull<ObjectType>, keys: *mut NonNull<KeyType>, count: NSUInteger, )

pub unsafe fn objectForKeyedSubscript( &self, key: &KeyType, ) -> Option<Retained<ObjectType>>

pub unsafe fn enumerateKeysAndObjectsUsingBlock( &self, block: &Block<dyn Fn(NonNull<KeyType>, NonNull<ObjectType>, NonNull<Bool>) + '_>, )

pub unsafe fn enumerateKeysAndObjectsWithOptions_usingBlock( &self, opts: NSEnumerationOptions, block: &Block<dyn Fn(NonNull<KeyType>, NonNull<ObjectType>, NonNull<Bool>) + '_>, )

pub unsafe fn keysSortedByValueUsingComparator( &self, cmptr: NSComparator, ) -> Retained<NSArray<KeyType>>

pub unsafe fn keysSortedByValueWithOptions_usingComparator( &self, opts: NSSortOptions, cmptr: NSComparator, ) -> Retained<NSArray<KeyType>>

pub unsafe fn keysOfEntriesPassingTest( &self, predicate: &Block<dyn Fn(NonNull<KeyType>, NonNull<ObjectType>, NonNull<Bool>) -> Bool + '_>, ) -> Retained<NSSet<KeyType>>

pub unsafe fn keysOfEntriesWithOptions_passingTest( &self, opts: NSEnumerationOptions, predicate: &Block<dyn Fn(NonNull<KeyType>, NonNull<ObjectType>, NonNull<Bool>) -> Bool + '_>, ) -> Retained<NSSet<KeyType>>

impl<KeyType: Message, ObjectType: Message> NSDictionary<KeyType, ObjectType>

NSDeprecated

pub unsafe fn getObjects_andKeys( &self, objects: *mut NonNull<ObjectType>, keys: *mut NonNull<KeyType>, )

👎Deprecated: Use -getObjects:andKeys:count: instead

pub unsafe fn dictionaryWithContentsOfFile( path: &NSString, ) -> Option<Retained<NSDictionary<KeyType, ObjectType>>>

👎Deprecated

pub unsafe fn dictionaryWithContentsOfURL( url: &NSURL, ) -> Option<Retained<NSDictionary<KeyType, ObjectType>>>

👎Deprecated

pub unsafe fn initWithContentsOfFile( this: Allocated<Self>, path: &NSString, ) -> Option<Retained<NSDictionary<KeyType, ObjectType>>>

👎Deprecated

pub unsafe fn initWithContentsOfURL( this: Allocated<Self>, url: &NSURL, ) -> Option<Retained<NSDictionary<KeyType, ObjectType>>>

👎Deprecated

pub unsafe fn writeToFile_atomically( &self, path: &NSString, use_auxiliary_file: bool, ) -> bool

👎Deprecated

pub unsafe fn writeToURL_atomically( &self, url: &NSURL, atomically: bool, ) -> bool

👎Deprecated

impl<KeyType: Message, ObjectType: Message> NSDictionary<KeyType, ObjectType>

NSDictionaryCreation

pub unsafe fn dictionary() -> Retained<Self>

pub unsafe fn dictionaryWithObject_forKey( object: &ObjectType, key: &ProtocolObject<dyn NSCopying>, ) -> Retained<Self>

pub unsafe fn dictionaryWithObjects_forKeys_count( objects: *mut NonNull<ObjectType>, keys: *mut NonNull<ProtocolObject<dyn NSCopying>>, cnt: NSUInteger, ) -> Retained<Self>

pub unsafe fn dictionaryWithDictionary( dict: &NSDictionary<KeyType, ObjectType>, ) -> Retained<Self>

pub unsafe fn dictionaryWithObjects_forKeys( objects: &NSArray<ObjectType>, keys: &NSArray<ProtocolObject<dyn NSCopying>>, ) -> Retained<Self>

pub unsafe fn initWithDictionary( this: Allocated<Self>, other_dictionary: &NSDictionary<KeyType, ObjectType>, ) -> Retained<Self>

pub unsafe fn initWithDictionary_copyItems( this: Allocated<Self>, other_dictionary: &NSDictionary<KeyType, ObjectType>, flag: bool, ) -> Retained<Self>

pub unsafe fn initWithObjects_forKeys( this: Allocated<Self>, objects: &NSArray<ObjectType>, keys: &NSArray<ProtocolObject<dyn NSCopying>>, ) -> Retained<Self>

impl<KeyType: Message, ObjectType: Message> NSDictionary<KeyType, ObjectType>

NSSharedKeySetDictionary

pub unsafe fn sharedKeySetForKeys( keys: &NSArray<ProtocolObject<dyn NSCopying>>, ) -> Retained<AnyObject>