objc2_foundation

Struct NSFileManager

pub struct NSFileManager { /* private fields */ }

Apple’s documentation

Implementations

impl NSFileManager

pub fn defaultManager() -> Retained<NSFileManager>

pub fn mountedVolumeURLsIncludingResourceValuesForKeys_options( &self, property_keys: Option<&NSArray<NSURLResourceKey>>, options: NSVolumeEnumerationOptions, ) -> Option<Retained<NSArray<NSURL>>>

pub unsafe fn unmountVolumeAtURL_options_completionHandler( &self, url: &NSURL, mask: NSFileManagerUnmountOptions, completion_handler: &DynBlock<dyn Fn(*mut NSError)>, )

Safety

completion_handler block must be sendable.

pub fn contentsOfDirectoryAtURL_includingPropertiesForKeys_options_error( &self, url: &NSURL, keys: Option<&NSArray<NSURLResourceKey>>, mask: NSDirectoryEnumerationOptions, ) -> Result<Retained<NSArray<NSURL>>, Retained<NSError>>

pub fn URLsForDirectory_inDomains( &self, directory: NSSearchPathDirectory, domain_mask: NSSearchPathDomainMask, ) -> Retained<NSArray<NSURL>>

pub fn URLForDirectory_inDomain_appropriateForURL_create_error( &self, directory: NSSearchPathDirectory, domain: NSSearchPathDomainMask, url: Option<&NSURL>, should_create: bool, ) -> Result<Retained<NSURL>, Retained<NSError>>

pub unsafe fn getRelationship_ofDirectoryAtURL_toItemAtURL_error( &self, out_relationship: NonNull<NSURLRelationship>, directory_url: &NSURL, other_url: &NSURL, ) -> Result<(), Retained<NSError>>

Safety

out_relationship must be a valid pointer.

pub unsafe fn getRelationship_ofDirectory_inDomain_toItemAtURL_error( &self, out_relationship: NonNull<NSURLRelationship>, directory: NSSearchPathDirectory, domain_mask: NSSearchPathDomainMask, url: &NSURL, ) -> Result<(), Retained<NSError>>

Safety

out_relationship must be a valid pointer.

pub unsafe fn createDirectoryAtURL_withIntermediateDirectories_attributes_error( &self, url: &NSURL, create_intermediates: bool, attributes: Option<&NSDictionary<NSFileAttributeKey, AnyObject>>, ) -> Result<(), Retained<NSError>>

Safety

attributes generic should be of the correct type.

pub fn createSymbolicLinkAtURL_withDestinationURL_error( &self, url: &NSURL, dest_url: &NSURL, ) -> Result<(), Retained<NSError>>

pub unsafe fn delegate( &self, ) -> Option<Retained<ProtocolObject<dyn NSFileManagerDelegate>>>

Safety

This is not retained internally, you must ensure the object is still alive.

pub unsafe fn setDelegate( &self, delegate: Option<&ProtocolObject<dyn NSFileManagerDelegate>>, )

Setter for delegate.

Safety

This is unretained, you must ensure the object is kept alive while in use.

pub unsafe fn setAttributes_ofItemAtPath_error( &self, attributes: &NSDictionary<NSFileAttributeKey, AnyObject>, path: &NSString, ) -> Result<(), Retained<NSError>>

Safety

attributes generic should be of the correct type.

pub unsafe fn createDirectoryAtPath_withIntermediateDirectories_attributes_error( &self, path: &NSString, create_intermediates: bool, attributes: Option<&NSDictionary<NSFileAttributeKey, AnyObject>>, ) -> Result<(), Retained<NSError>>

Safety

attributes generic should be of the correct type.

pub fn contentsOfDirectoryAtPath_error( &self, path: &NSString, ) -> Result<Retained<NSArray<NSString>>, Retained<NSError>>

pub fn subpathsOfDirectoryAtPath_error( &self, path: &NSString, ) -> Result<Retained<NSArray<NSString>>, Retained<NSError>>

pub fn attributesOfItemAtPath_error( &self, path: &NSString, ) -> Result<Retained<NSDictionary<NSFileAttributeKey, AnyObject>>, Retained<NSError>>

pub fn attributesOfFileSystemForPath_error( &self, path: &NSString, ) -> Result<Retained<NSDictionary<NSFileAttributeKey, AnyObject>>, Retained<NSError>>

pub fn createSymbolicLinkAtPath_withDestinationPath_error( &self, path: &NSString, dest_path: &NSString, ) -> Result<(), Retained<NSError>>

pub fn destinationOfSymbolicLinkAtPath_error( &self, path: &NSString, ) -> Result<Retained<NSString>, Retained<NSError>>

pub fn copyItemAtPath_toPath_error( &self, src_path: &NSString, dst_path: &NSString, ) -> Result<(), Retained<NSError>>

pub fn moveItemAtPath_toPath_error( &self, src_path: &NSString, dst_path: &NSString, ) -> Result<(), Retained<NSError>>

pub fn linkItemAtPath_toPath_error( &self, src_path: &NSString, dst_path: &NSString, ) -> Result<(), Retained<NSError>>

pub fn removeItemAtPath_error( &self, path: &NSString, ) -> Result<(), Retained<NSError>>

pub fn copyItemAtURL_toURL_error( &self, src_url: &NSURL, dst_url: &NSURL, ) -> Result<(), Retained<NSError>>

pub fn moveItemAtURL_toURL_error( &self, src_url: &NSURL, dst_url: &NSURL, ) -> Result<(), Retained<NSError>>

pub fn linkItemAtURL_toURL_error( &self, src_url: &NSURL, dst_url: &NSURL, ) -> Result<(), Retained<NSError>>

pub fn removeItemAtURL_error( &self, url: &NSURL, ) -> Result<(), Retained<NSError>>

pub fn trashItemAtURL_resultingItemURL_error( &self, url: &NSURL, out_resulting_url: Option<&mut Option<Retained<NSURL>>>, ) -> Result<(), Retained<NSError>>

pub fn fileAttributesAtPath_traverseLink( &self, path: &NSString, yorn: bool, ) -> Option<Retained<NSDictionary>>

👎Deprecated: Use -attributesOfItemAtPath:error: instead

pub unsafe fn changeFileAttributes_atPath( &self, attributes: &NSDictionary, path: &NSString, ) -> bool

👎Deprecated: Use -setAttributes:ofItemAtPath:error: instead

Safety

attributes generic should be of the correct type.

pub fn directoryContentsAtPath( &self, path: &NSString, ) -> Option<Retained<NSArray>>

👎Deprecated: Use -contentsOfDirectoryAtPath:error: instead

pub fn fileSystemAttributesAtPath( &self, path: &NSString, ) -> Option<Retained<NSDictionary>>

👎Deprecated: Use -attributesOfFileSystemForPath:error: instead

pub fn pathContentOfSymbolicLinkAtPath( &self, path: &NSString, ) -> Option<Retained<NSString>>

👎Deprecated: Use -destinationOfSymbolicLinkAtPath:error:

pub fn createSymbolicLinkAtPath_pathContent( &self, path: &NSString, otherpath: &NSString, ) -> bool

👎Deprecated: Use -createSymbolicLinkAtPath:error: instead

pub unsafe fn createDirectoryAtPath_attributes( &self, path: &NSString, attributes: &NSDictionary, ) -> bool

👎Deprecated: Use -createDirectoryAtPath:withIntermediateDirectories:attributes:error: instead

Safety

attributes generic should be of the correct type.

pub unsafe fn linkPath_toPath_handler( &self, src: &NSString, dest: &NSString, handler: Option<&AnyObject>, ) -> bool

👎Deprecated: Not supported

Safety

handler should be of the correct type.

pub unsafe fn copyPath_toPath_handler( &self, src: &NSString, dest: &NSString, handler: Option<&AnyObject>, ) -> bool

👎Deprecated: Not supported

Safety

handler should be of the correct type.

pub unsafe fn movePath_toPath_handler( &self, src: &NSString, dest: &NSString, handler: Option<&AnyObject>, ) -> bool

👎Deprecated: Not supported

Safety

handler should be of the correct type.

pub unsafe fn removeFileAtPath_handler( &self, path: &NSString, handler: Option<&AnyObject>, ) -> bool

👎Deprecated: Not supported

Safety

handler should be of the correct type.

pub fn currentDirectoryPath(&self) -> Retained<NSString>

pub fn changeCurrentDirectoryPath(&self, path: &NSString) -> bool

pub fn fileExistsAtPath(&self, path: &NSString) -> bool

pub unsafe fn fileExistsAtPath_isDirectory( &self, path: &NSString, is_directory: *mut Bool, ) -> bool

Safety

is_directory must be a valid pointer or null.

pub fn isReadableFileAtPath(&self, path: &NSString) -> bool

pub fn isWritableFileAtPath(&self, path: &NSString) -> bool

pub fn isExecutableFileAtPath(&self, path: &NSString) -> bool

pub fn isDeletableFileAtPath(&self, path: &NSString) -> bool

pub fn contentsEqualAtPath_andPath( &self, path1: &NSString, path2: &NSString, ) -> bool

pub fn displayNameAtPath(&self, path: &NSString) -> Retained<NSString>

pub fn componentsToDisplayForPath( &self, path: &NSString, ) -> Option<Retained<NSArray<NSString>>>

pub fn enumeratorAtPath( &self, path: &NSString, ) -> Option<Retained<NSDirectoryEnumerator<NSString>>>

pub fn enumeratorAtURL_includingPropertiesForKeys_options_errorHandler( &self, url: &NSURL, keys: Option<&NSArray<NSURLResourceKey>>, mask: NSDirectoryEnumerationOptions, handler: Option<&DynBlock<dyn Fn(NonNull<NSURL>, NonNull<NSError>) -> Bool>>, ) -> Option<Retained<NSDirectoryEnumerator<NSURL>>>

pub fn subpathsAtPath( &self, path: &NSString, ) -> Option<Retained<NSArray<NSString>>>

pub fn contentsAtPath(&self, path: &NSString) -> Option<Retained<NSData>>

pub unsafe fn createFileAtPath_contents_attributes( &self, path: &NSString, data: Option<&NSData>, attr: Option<&NSDictionary<NSFileAttributeKey, AnyObject>>, ) -> bool

Safety

attr generic should be of the correct type.

pub fn fileSystemRepresentationWithPath( &self, path: &NSString, ) -> NonNull<c_char>

pub unsafe fn stringWithFileSystemRepresentation_length( &self, str: NonNull<c_char>, len: NSUInteger, ) -> Retained<NSString>

Safety

str must be a valid pointer.

pub fn replaceItemAtURL_withItemAtURL_backupItemName_options_resultingItemURL_error( &self, original_item_url: &NSURL, new_item_url: &NSURL, backup_item_name: Option<&NSString>, options: NSFileManagerItemReplacementOptions, resulting_url: Option<&mut Option<Retained<NSURL>>>, ) -> Result<(), Retained<NSError>>

pub fn setUbiquitous_itemAtURL_destinationURL_error( &self, flag: bool, url: &NSURL, destination_url: &NSURL, ) -> Result<(), Retained<NSError>>

pub fn isUbiquitousItemAtURL(&self, url: &NSURL) -> bool

pub fn startDownloadingUbiquitousItemAtURL_error( &self, url: &NSURL, ) -> Result<(), Retained<NSError>>

pub fn evictUbiquitousItemAtURL_error( &self, url: &NSURL, ) -> Result<(), Retained<NSError>>

pub fn URLForUbiquityContainerIdentifier( &self, container_identifier: Option<&NSString>, ) -> Option<Retained<NSURL>>

pub fn URLForPublishingUbiquitousItemAtURL_expirationDate_error( &self, url: &NSURL, out_date: Option<&mut Option<Retained<NSDate>>>, ) -> Result<Retained<NSURL>, Retained<NSError>>

pub fn ubiquityIdentityToken(&self) -> Option<Retained<AnyObject>>

pub unsafe fn pauseSyncForUbiquitousItemAtURL_completionHandler( &self, url: &NSURL, completion_handler: &DynBlock<dyn Fn(*mut NSError)>, )

Asynchronously pauses sync of an item at the given URL.

Call this when opening an item to prevent sync from altering the contents of the URL. Once paused, the file provider will not upload local changes nor download remote changes.

While paused, call uploadLocalVersionOfUbiquitousItem(at:withConflictResolutionPolicy:completionHandler:) when the document is in a stable state. This action keeps the server version as up-to-date as possible.

If the item is already paused, a second call to this method reports success. If the file provider is already applying changes to the item, the pause fails with an NSFileWriteUnknownError-enum.case, with an underlying error that has domain NSPOSIXErrorDomain and code POSIXError/EBUSY. If the pause fails, wait for the state to stabilize before retrying. Pausing also fails with CocoaError/featureUnsupported if url refers to a regular (non-package) directory.

Pausing sync is independent of the calling app’s lifecycle; sync doesn’t automatically resume if the app closes or crashes and relaunches later. To resume syncing, explicitly call resumeSyncForUbiquitousItem(at:with:completionHandler:). Always be sure to resume syncing before you close the item.

Parameters:

url: The URL of the item for which to pause sync.

completionHandler: A closure or block that the framework calls when the pause action completes. It receives a single NSError parameter to indicate an error that prevented pausing; this value is nil if the pause succeeded. In Swift, you can omit the completion handler and catch the thrown error instead.

Safety

completion_handler block must be sendable.

pub unsafe fn resumeSyncForUbiquitousItemAtURL_withBehavior_completionHandler( &self, url: &NSURL, behavior: NSFileManagerResumeSyncBehavior, completion_handler: &DynBlock<dyn Fn(*mut NSError)>, )

Asynchronously resumes the sync on a paused item using the given resume behavior.

Always call this method when your app closes an item to allow the file provider to sync local changes back to the server.

In most situations, the NSFileManagerResumeSyncBehavior/preserveLocalChanges behavior is the best choice to avoid any risk of data loss.

The resume call fails with CocoaError/featureUnsupported if url isn’t currently paused. If the device isn’t connected to the network, the call may fail with NSFileWriteUnknownError-enum.case, with the underlying error of <doc ://com.apple.documentation/documentation/FileProvider/NSFileProviderError/serverUnreachable>.

Parameters:

url: The URL of the item for which to resume sync.

behavior: A NSFileManagerResumeSyncBehavior value that tells the file manager how to handle conflicts between local and remote versions of files.

completionHandler: A closure or block that the framework calls when the resume action completes. It receives a single NSError parameter to indicate an error that prevented the resume action; the value is nil if the resume succeeded. In Swift, you can omit the completion handler and catch the thrown error instead.

Safety

completion_handler block must be sendable.

pub unsafe fn fetchLatestRemoteVersionOfItemAtURL_completionHandler( &self, url: &NSURL, completion_handler: &DynBlock<dyn Fn(*mut NSFileVersion, *mut NSError)>, )

Asynchronously fetches the latest remote version of a given item from the server.

Use this method if uploading fails due to a version conflict and sync is paused. In this case, fetching the latest remote version allows you to inspect the newer item from the server, resolve the conflict, and resume uploading.

The version provided by this call depends on several factors:

If there is no newer version of the file on the server, the caller receives the current version of the file.

If the server has a newer version and sync isn’t paused, this call replaces the local item and provides the version of the new item.

If the server has a newer version but sync is paused, the returned version points to a side location. In this case, call NSFileVersion/replaceItem(at:options:) on the provided version object to replace the local item with the newer item from the server.

If the device isn’t connected to the network, the call may fail with NSFileReadUnknownError-enum.case, with the underlying error of <doc ://com.apple.documentation/documentation/FileProvider/NSFileProviderError/serverUnreachable>.

Parameters:

url: The URL of the item for which to check the version.

completionHandler: A closure or block that the framework calls when the fetch action completes. It receives parameters of types NSFileVersion and NSError. The error is nil if fetching the remote version succeeded; otherwise it indicates the error that caused the call to fail. In Swift, you can omit the completion handler, catching any error in a do-catch block and receiving the version as the return value.

Safety

completion_handler block must be sendable.

pub unsafe fn uploadLocalVersionOfUbiquitousItemAtURL_withConflictResolutionPolicy_completionHandler( &self, url: &NSURL, conflict_resolution_policy: NSFileManagerUploadLocalVersionConflictPolicy, completion_handler: &DynBlock<dyn Fn(*mut NSFileVersion, *mut NSError)>, )

Asynchronously uploads the local version of the item using the provided conflict resolution policy.

Once your app pauses a sync for an item, call this method every time your document is in a stable state. This action keeps the server version as up-to-date as possible.

If the server has a newer version than the one to which the app made changes, uploading fails with NSFileWriteUnknownError-enum.case, with an underlying error of <doc ://com.apple.documentation/documentation/FileProvider/NSFileProviderError/localVersionConflictingWithServer>. In this case, call FileManager/fetchLatestRemoteVersionOfItem(at:completionHandler:), rebase local changes on top of that version, and retry the upload.

If the device isn’t connected to the network, the call may fail with NSFileWriteUnknownError-enum.case, with the underlying error of <doc ://com.apple.documentation/documentation/FileProvider/NSFileProviderError/serverUnreachable>.

Parameters:

url: The URL of the item for which to check the version.

conflictResolutionPolicy: The policy the file manager applies if the local and server versions conflict.

completionHandler: A closure or block that the framework calls when the upload completes. It receives parameters of types NSFileVersion and NSError. The error is nil if fetching the remote version succeeded; otherwise it indicates the error that caused the call to fail. In Swift, you can omit the completion handler, catching any error in a do-catch block and receiving the version as the return value.

Safety

completion_handler block must be sendable.

pub unsafe fn getFileProviderServicesForItemAtURL_completionHandler( &self, url: &NSURL, completion_handler: &DynBlock<dyn Fn(*mut NSDictionary<NSFileProviderServiceName, NSFileProviderService>, *mut NSError)>, )

Safety

completion_handler block must be sendable.

pub fn containerURLForSecurityApplicationGroupIdentifier( &self, group_identifier: &NSString, ) -> Option<Retained<NSURL>>

impl NSFileManager

Methods declared on superclass NSObject.

pub fn init(this: Allocated<Self>) -> Retained<Self>

pub fn new() -> Retained<Self>

impl NSFileManager

NSUserInformation.

pub fn homeDirectoryForCurrentUser(&self) -> Retained<NSURL>

pub fn temporaryDirectory(&self) -> Retained<NSURL>

pub fn homeDirectoryForUser( &self, user_name: &NSString, ) -> Option<Retained<NSURL>>