objc2_foundation

Struct NSMutableArray

pub struct NSMutableArray<ObjectType: ?Sized = AnyObject> { /* private fields */ }

************** Mutable Array ***************

See also Apple’s documentation

Implementations

impl<ObjectType: Message> NSMutableArray<ObjectType>

Convenience creation methods.

pub fn from_slice(slice: &[&ObjectType]) -> Retained<Self>

pub fn from_retained_slice(slice: &[Retained<ObjectType>]) -> Retained<Self>

impl<ObjectType: Message> NSMutableArray<ObjectType>

Convenience mutation methods.

pub fn insert(&self, index: usize, obj: &ObjectType)

Insert an object into the array at the given index.

Panics

Panics if the index is out of bounds.

pub fn sort_by<F: FnMut(&ObjectType, &ObjectType) -> Ordering>( &self, compare: F, )

Sort the array by the given comparison closure.

impl<ObjectType: Message> NSMutableArray<ObjectType>

Methods declared on superclass NSArray.

NSArrayCreation.

pub fn array() -> Retained<Self>

pub fn arrayWithObject(an_object: &ObjectType) -> Retained<Self>

pub unsafe fn arrayWithObjects_count( objects: NonNull<NonNull<ObjectType>>, cnt: NSUInteger, ) -> Retained<Self>

Safety

objects must be a valid pointer.

pub fn arrayWithArray(array: &NSArray<ObjectType>) -> Retained<Self>

pub fn initWithArray( this: Allocated<Self>, array: &NSArray<ObjectType>, ) -> Retained<Self>

pub unsafe fn initWithArray_copyItems( this: Allocated<Self>, array: &NSArray<ObjectType>, flag: bool, ) -> Retained<Self>

impl<ObjectType: ?Sized + Message> NSMutableArray<ObjectType>

pub unsafe fn cast_unchecked<NewObjectType: ?Sized + Message>( &self, ) -> &NSMutableArray<NewObjectType>

Unchecked conversion of the generic parameter.

Safety

The generic must be valid to reinterpret as the given type.

impl<ObjectType: Message> NSMutableArray<ObjectType>

pub fn addObject(&self, an_object: &ObjectType)

pub fn insertObject_atIndex(&self, an_object: &ObjectType, index: NSUInteger)

pub fn removeLastObject(&self)

pub fn removeObjectAtIndex(&self, index: NSUInteger)

pub fn replaceObjectAtIndex_withObject( &self, index: NSUInteger, an_object: &ObjectType, )

pub fn init(this: Allocated<Self>) -> Retained<Self>

pub fn initWithCapacity( this: Allocated<Self>, num_items: NSUInteger, ) -> Retained<Self>

pub unsafe fn initWithCoder( this: Allocated<Self>, coder: &NSCoder, ) -> Option<Retained<Self>>

Safety

coder possibly has further requirements.

impl<ObjectType: Message> NSMutableArray<ObjectType>

Methods declared on superclass NSArray.

pub unsafe fn initWithObjects_count( this: Allocated<Self>, objects: *mut NonNull<ObjectType>, cnt: NSUInteger, ) -> Retained<Self>

Safety

objects must be a valid pointer or null.

impl<ObjectType: Message> NSMutableArray<ObjectType>

Methods declared on superclass NSObject.

pub fn new() -> Retained<Self>

impl<ObjectType: Message> NSMutableArray<ObjectType>

NSExtendedMutableArray.

pub fn addObjectsFromArray(&self, other_array: &NSArray<ObjectType>)

pub fn exchangeObjectAtIndex_withObjectAtIndex( &self, idx1: NSUInteger, idx2: NSUInteger, )

pub fn removeAllObjects(&self)

pub fn removeObject_inRange(&self, an_object: &ObjectType, range: NSRange)

pub fn removeObject(&self, an_object: &ObjectType)

pub fn removeObjectIdenticalTo_inRange( &self, an_object: &ObjectType, range: NSRange, )

pub fn removeObjectIdenticalTo(&self, an_object: &ObjectType)

pub unsafe fn removeObjectsFromIndices_numIndices( &self, indices: NonNull<NSUInteger>, cnt: NSUInteger, )

👎Deprecated: Not supported

Safety

indices must be a valid pointer.

pub fn removeObjectsInArray(&self, other_array: &NSArray<ObjectType>)

pub fn removeObjectsInRange(&self, range: NSRange)

pub fn replaceObjectsInRange_withObjectsFromArray_range( &self, range: NSRange, other_array: &NSArray<ObjectType>, other_range: NSRange, )

pub fn replaceObjectsInRange_withObjectsFromArray( &self, range: NSRange, other_array: &NSArray<ObjectType>, )

pub fn setArray(&self, other_array: &NSArray<ObjectType>)

pub unsafe fn sortUsingFunction_context( &self, compare: unsafe extern "C-unwind" fn(NonNull<ObjectType>, NonNull<ObjectType>, *mut c_void) -> NSInteger, context: *mut c_void, )

Safety

compare must be implemented correctly.

context must be a valid pointer or null.

pub unsafe fn sortUsingSelector(&self, comparator: Sel)

Safety

comparator must be a valid selector.

pub fn insertObjects_atIndexes( &self, objects: &NSArray<ObjectType>, indexes: &NSIndexSet, )

pub fn removeObjectsAtIndexes(&self, indexes: &NSIndexSet)

pub fn replaceObjectsAtIndexes_withObjects( &self, indexes: &NSIndexSet, objects: &NSArray<ObjectType>, )

pub fn setObject_atIndexedSubscript(&self, obj: &ObjectType, idx: NSUInteger)

pub unsafe fn sortUsingComparator(&self, cmptr: NSComparator)

Safety

cmptr must be a valid pointer.

pub unsafe fn sortWithOptions_usingComparator( &self, opts: NSSortOptions, cmptr: NSComparator, )

Safety

cmptr must be a valid pointer.

impl<ObjectType: Message> NSMutableArray<ObjectType>

NSMutableArrayCreation.

pub fn arrayWithCapacity(num_items: NSUInteger) -> Retained<Self>

pub unsafe fn arrayWithContentsOfFile( path: &NSString, ) -> Option<Retained<NSMutableArray<ObjectType>>>

pub unsafe fn arrayWithContentsOfURL( url: &NSURL, ) -> Option<Retained<NSMutableArray<ObjectType>>>

pub unsafe fn initWithContentsOfFile( this: Allocated<Self>, path: &NSString, ) -> Option<Retained<NSMutableArray<ObjectType>>>

pub unsafe fn initWithContentsOfURL( this: Allocated<Self>, url: &NSURL, ) -> Option<Retained<NSMutableArray<ObjectType>>>

impl<ObjectType: Message> NSMutableArray<ObjectType>

NSMutableArrayDiffing.

pub fn applyDifference( &self, difference: &NSOrderedCollectionDifference<ObjectType>, )

impl<ObjectType: Message> NSMutableArray<ObjectType>

NSPredicateSupport.

pub fn filterUsingPredicate(&self, predicate: &NSPredicate)

impl<ObjectType: Message> NSMutableArray<ObjectType>

NSSortDescriptorSorting.

pub fn sortUsingDescriptors(&self, sort_descriptors: &NSArray<NSSortDescriptor>)