objc2_foundation

Struct NSMutableData

#[repr(C)]
pub struct NSMutableData { /* private fields */ }

************** Mutable Data ***************

See also Apple’s documentation

Implementations

impl NSMutableData

pub fn with_bytes(bytes: &[u8]) -> Retained<Self>

pub fn from_vec(bytes: Vec<u8>) -> Retained<Self>

impl NSMutableData

pub unsafe fn as_mut_bytes_unchecked(&self) -> &mut [u8]

A mutable view of the bytes in the data.

Safety

No methods on the NSMutableData may be called while the returned slice is alive.

pub fn extend_from_slice(&self, bytes: &[u8])

pub fn push(&self, byte: u8)

pub fn replace_range(&self, range: Range<usize>, bytes: &[u8])

pub fn set_bytes(&self, bytes: &[u8])

impl NSMutableData

Methods declared on superclass NSData.

NSDataCreation.

pub unsafe fn data() -> Retained<Self>

pub unsafe fn dataWithBytes_length( bytes: *const c_void, length: NSUInteger, ) -> Retained<Self>

pub unsafe fn dataWithBytesNoCopy_length( bytes: NonNull<c_void>, length: NSUInteger, ) -> Retained<Self>

pub unsafe fn dataWithBytesNoCopy_length_freeWhenDone( bytes: NonNull<c_void>, length: NSUInteger, b: bool, ) -> Retained<Self>

pub unsafe fn dataWithContentsOfFile_options_error( path: &NSString, read_options_mask: NSDataReadingOptions, ) -> Result<Retained<Self>, Retained<NSError>>

pub unsafe fn dataWithContentsOfURL_options_error( url: &NSURL, read_options_mask: NSDataReadingOptions, ) -> Result<Retained<Self>, Retained<NSError>>

pub unsafe fn dataWithContentsOfFile(path: &NSString) -> Option<Retained<Self>>

pub unsafe fn dataWithContentsOfURL(url: &NSURL) -> Option<Retained<Self>>

pub unsafe fn initWithBytes_length( this: Allocated<Self>, bytes: *const c_void, length: NSUInteger, ) -> Retained<Self>

pub unsafe fn initWithBytesNoCopy_length( this: Allocated<Self>, bytes: NonNull<c_void>, length: NSUInteger, ) -> Retained<Self>

pub unsafe fn initWithBytesNoCopy_length_freeWhenDone( this: Allocated<Self>, bytes: NonNull<c_void>, length: NSUInteger, b: bool, ) -> Retained<Self>

pub unsafe fn initWithBytesNoCopy_length_deallocator( this: Allocated<Self>, bytes: NonNull<c_void>, length: NSUInteger, deallocator: Option<&Block<dyn Fn(NonNull<c_void>, NSUInteger)>>, ) -> Retained<Self>

pub unsafe fn initWithContentsOfFile_options_error( this: Allocated<Self>, path: &NSString, read_options_mask: NSDataReadingOptions, ) -> Result<Retained<Self>, Retained<NSError>>

pub unsafe fn initWithContentsOfURL_options_error( this: Allocated<Self>, url: &NSURL, read_options_mask: NSDataReadingOptions, ) -> Result<Retained<Self>, Retained<NSError>>

pub unsafe fn initWithContentsOfFile( this: Allocated<Self>, path: &NSString, ) -> Option<Retained<Self>>

pub unsafe fn initWithContentsOfURL( this: Allocated<Self>, url: &NSURL, ) -> Option<Retained<Self>>

pub unsafe fn initWithData( this: Allocated<Self>, data: &NSData, ) -> Retained<Self>

pub fn dataWithData(data: &NSData) -> Retained<Self>

impl NSMutableData

Methods declared on superclass NSData.

NSDataBase64Encoding.

pub unsafe fn initWithBase64EncodedString_options( this: Allocated<Self>, base64_string: &NSString, options: NSDataBase64DecodingOptions, ) -> Option<Retained<Self>>

pub unsafe fn initWithBase64EncodedData_options( this: Allocated<Self>, base64_data: &NSData, options: NSDataBase64DecodingOptions, ) -> Option<Retained<Self>>

impl NSMutableData

Methods declared on superclass NSData.

NSDeprecated.

pub unsafe fn initWithContentsOfMappedFile( this: Allocated<Self>, path: &NSString, ) -> Option<Retained<Self>>

👎Deprecated: Use -initWithContentsOfURL:options:error: and NSDataReadingMappedIfSafe or NSDataReadingMappedAlways instead.

pub unsafe fn initWithBase64Encoding( this: Allocated<Self>, base64_string: &NSString, ) -> Option<Retained<Self>>

👎Deprecated: Use initWithBase64EncodedString:options: instead

impl NSMutableData

pub fn setLength(&self, length: NSUInteger)

Setter for [length][Self::length].

impl NSMutableData

Methods declared on superclass NSObject.

pub fn init(this: Allocated<Self>) -> Retained<Self>

pub fn new() -> Retained<Self>

impl NSMutableData

NSExtendedMutableData.

pub unsafe fn appendBytes_length( &self, bytes: NonNull<c_void>, length: NSUInteger, )

pub unsafe fn appendData(&self, other: &NSData)

pub unsafe fn increaseLengthBy(&self, extra_length: NSUInteger)

pub unsafe fn replaceBytesInRange_withBytes( &self, range: NSRange, bytes: NonNull<c_void>, )

pub unsafe fn resetBytesInRange(&self, range: NSRange)

pub unsafe fn setData(&self, data: &NSData)

pub unsafe fn replaceBytesInRange_withBytes_length( &self, range: NSRange, replacement_bytes: *const c_void, replacement_length: NSUInteger, )

impl NSMutableData

NSMutableDataCreation.

pub fn dataWithCapacity(a_num_items: NSUInteger) -> Option<Retained<Self>>

pub unsafe fn dataWithLength(length: NSUInteger) -> Option<Retained<Self>>

pub fn initWithCapacity( this: Allocated<Self>, capacity: NSUInteger, ) -> Option<Retained<Self>>

pub unsafe fn initWithLength( this: Allocated<Self>, length: NSUInteger, ) -> Option<Retained<Self>>

impl NSMutableData

NSMutableDataCompression.

pub unsafe fn decompressUsingAlgorithm_error( &self, algorithm: NSDataCompressionAlgorithm, ) -> Result<(), Retained<NSError>>

pub unsafe fn compressUsingAlgorithm_error( &self, algorithm: NSDataCompressionAlgorithm, ) -> Result<(), Retained<NSError>>