objc2_foundation

Struct NSMutableDictionary

#[repr(C)]
pub struct NSMutableDictionary<KeyType: ?Sized = AnyObject, ObjectType: ?Sized = AnyObject> { /* private fields */ }

************** Mutable Dictionary ***************

See also Apple’s documentation

Implementations

impl<KeyType: Message, ObjectType: Message> NSMutableDictionary<KeyType, ObjectType>

Convenience creation methods.

pub fn from_slices<CopiedKey>( keys: &[&CopiedKey], objects: &[&ObjectType], ) -> Retained<Self>
where CopiedKey: Message + NSCopying + CopyingHelper<Result = KeyType>,

pub fn from_retained_objects<CopiedKey>( keys: &[&CopiedKey], objects: &[Retained<ObjectType>], ) -> Retained<Self>
where CopiedKey: Message + NSCopying + CopyingHelper<Result = KeyType>,

impl<KeyType: Message, ObjectType: Message> NSMutableDictionary<KeyType, ObjectType>

Convenience mutation methods.

pub fn insert<CopiedKey>(&self, key: &CopiedKey, object: &ObjectType)
where CopiedKey: Message + NSCopying + CopyingHelper<Result = KeyType>,

Inserts a key-value pair into the dictionary.

If the dictionary did not have this key present, the value is inserted. If the dictionary already had this key present, the value and the key is updated.

Examples

use objc2_foundation::{ns_string, NSMutableDictionary, NSObject};

let dict = NSMutableDictionary::new();
dict.insert(ns_string!("key"), &*NSObject::new());

impl<KeyType: Message, ObjectType: Message> NSMutableDictionary<KeyType, ObjectType>

Methods declared on superclass NSDictionary.

NSDictionaryCreation.

pub unsafe fn dictionary() -> Retained<Self>

pub unsafe fn dictionaryWithObject_forKey( object: &ObjectType, key: &ProtocolObject<dyn NSCopying>, ) -> Retained<Self>

pub unsafe fn dictionaryWithObjects_forKeys_count( objects: *mut NonNull<ObjectType>, keys: *mut NonNull<ProtocolObject<dyn NSCopying>>, cnt: NSUInteger, ) -> Retained<Self>

pub unsafe fn dictionaryWithDictionary( dict: &NSDictionary<KeyType, ObjectType>, ) -> Retained<Self>

pub unsafe fn dictionaryWithObjects_forKeys( objects: &NSArray<ObjectType>, keys: &NSArray<ProtocolObject<dyn NSCopying>>, ) -> Retained<Self>

pub unsafe fn initWithDictionary( this: Allocated<Self>, other_dictionary: &NSDictionary<KeyType, ObjectType>, ) -> Retained<Self>

pub unsafe fn initWithDictionary_copyItems( this: Allocated<Self>, other_dictionary: &NSDictionary<KeyType, ObjectType>, flag: bool, ) -> Retained<Self>

pub unsafe fn initWithObjects_forKeys( this: Allocated<Self>, objects: &NSArray<ObjectType>, keys: &NSArray<ProtocolObject<dyn NSCopying>>, ) -> Retained<Self>

impl<KeyType: Message, ObjectType: Message> NSMutableDictionary<KeyType, ObjectType>

pub fn removeObjectForKey(&self, a_key: &KeyType)

pub unsafe fn setObject_forKey( &self, an_object: &ObjectType, a_key: &ProtocolObject<dyn NSCopying>, )

pub fn init(this: Allocated<Self>) -> Retained<Self>

pub fn initWithCapacity( this: Allocated<Self>, num_items: NSUInteger, ) -> Retained<Self>

pub unsafe fn initWithCoder( this: Allocated<Self>, coder: &NSCoder, ) -> Option<Retained<Self>>

impl<KeyType: Message, ObjectType: Message> NSMutableDictionary<KeyType, ObjectType>

Methods declared on superclass NSDictionary.

pub unsafe fn initWithObjects_forKeys_count( this: Allocated<Self>, objects: *mut NonNull<ObjectType>, keys: *mut NonNull<ProtocolObject<dyn NSCopying>>, cnt: NSUInteger, ) -> Retained<Self>

impl<KeyType: Message, ObjectType: Message> NSMutableDictionary<KeyType, ObjectType>

Methods declared on superclass NSObject.

pub fn new() -> Retained<Self>

impl<KeyType: Message, ObjectType: Message> NSMutableDictionary<KeyType, ObjectType>

NSExtendedMutableDictionary.

pub unsafe fn addEntriesFromDictionary( &self, other_dictionary: &NSDictionary<KeyType, ObjectType>, )

pub fn removeAllObjects(&self)

pub unsafe fn removeObjectsForKeys(&self, key_array: &NSArray<KeyType>)

pub unsafe fn setDictionary( &self, other_dictionary: &NSDictionary<KeyType, ObjectType>, )

pub unsafe fn setObject_forKeyedSubscript( &self, obj: Option<&ObjectType>, key: &ProtocolObject<dyn NSCopying>, )

impl<KeyType: Message, ObjectType: Message> NSMutableDictionary<KeyType, ObjectType>

NSMutableDictionaryCreation.

pub unsafe fn dictionaryWithCapacity(num_items: NSUInteger) -> Retained<Self>

pub unsafe fn dictionaryWithContentsOfFile( path: &NSString, ) -> Option<Retained<NSMutableDictionary<KeyType, ObjectType>>>

pub unsafe fn dictionaryWithContentsOfURL( url: &NSURL, ) -> Option<Retained<NSMutableDictionary<KeyType, ObjectType>>>

pub unsafe fn initWithContentsOfFile( this: Allocated<Self>, path: &NSString, ) -> Option<Retained<NSMutableDictionary<KeyType, ObjectType>>>

pub unsafe fn initWithContentsOfURL( this: Allocated<Self>, url: &NSURL, ) -> Option<Retained<NSMutableDictionary<KeyType, ObjectType>>>

impl<KeyType: Message, ObjectType: Message> NSMutableDictionary<KeyType, ObjectType>

NSSharedKeySetDictionary.

pub unsafe fn dictionaryWithSharedKeySet( keyset: &AnyObject, ) -> Retained<NSMutableDictionary<KeyType, ObjectType>>

impl<KeyType: Message, ObjectType: Message> NSMutableDictionary<KeyType, ObjectType>

NSKeyValueCoding.

pub unsafe fn setValue_forKey(&self, value: Option<&ObjectType>, key: &NSString)