objc2_foundation

Struct NSUndoManager

pub struct NSUndoManager { /* private fields */ }

Apple’s documentation

Implementations

impl NSUndoManager

pub fn beginUndoGrouping(&self)

Marks the beginning of an undo group.

All individual undo operations before a subsequent endUndoGrouping message are grouped together and reversed by a later undo message. By default undo groups are begun automatically at the start of the event loop, but you can begin your own undo groups with this method, and nest them within other groups.

This method posts an NSUndoManagerCheckpointNotification unless a top-level undo is in progress. It posts an NSUndoManagerDidOpenUndoGroupNotification if a new group was successfully created.

pub fn endUndoGrouping(&self)

Marks the end of an undo group.

All individual undo operations back to the matching beginUndoGrouping message are grouped together and reversed by a later undo or undoNestedGroup message. Undo groups can be nested, thus providing functionality similar to nested transactions. Raises an NSInternalInconsistencyException if there’s no beginUndoGrouping message in effect.

This method posts an NSUndoManagerCheckpointNotification and an NSUndoManagerDidCloseUndoGroupNotification just before the group is closed.

pub fn groupingLevel(&self) -> NSInteger

The number of nested undo groups (or redo groups, if Redo was invoked last) in the current event loop.

An integer indicating the number of nested groups. If 0 is returned, there is no open undo or redo group.

pub fn disableUndoRegistration(&self)

Disables the recording of undo operations, whether by registerUndoWithTarget:selector:object: or by invocation-based undo.

This method can be invoked multiple times by multiple clients. The enableUndoRegistration method must be invoked an equal number of times to re-enable undo registration.

pub fn enableUndoRegistration(&self)

Enables the recording of undo operations.

Because undo registration is enabled by default, this is used to balance a prior disableUndoRegistration. Undo registration isn’t actually re-enabled until an enable message balances the last disable message in effect. Raises an NSInternalInconsistencyException if invoked while no disableUndoRegistration() message is in effect.

pub fn isUndoRegistrationEnabled(&self) -> bool

Whether the recording of undo operations is enabled.

pub fn groupsByEvent(&self) -> bool

A Boolean value that indicates whether the receiver automatically creates undo groups around each pass of the run loop.

If true, the receiver automatically creates undo groups around each pass of the run loop. The default is true. If you turn automatic grouping off, you must close groups explicitly before invoking either undo or undoNestedGroup.

pub fn setGroupsByEvent(&self, groups_by_event: bool)

Setter for groupsByEvent.

pub fn levelsOfUndo(&self) -> NSUInteger

The maximum number of top-level undo groups the receiver holds.

An integer specifying the number of undo groups. A limit of 0 indicates no limit, so old undo groups are never dropped. When ending an undo group results in the number of groups exceeding this limit, the oldest groups are dropped from the stack. The default is 0. If you change the limit to a level below the prior limit, old undo groups are immediately dropped.

pub fn setLevelsOfUndo(&self, levels_of_undo: NSUInteger)

Setter for levelsOfUndo.

pub fn runLoopModes(&self) -> Retained<NSArray<NSRunLoopMode>>

The modes governing the types of input handled during a cycle of the run loop.

An array of string constants specifying the current run-loop modes. By default, the sole run-loop mode is NSDefaultRunLoopMode (which excludes data from NSConnection objects). Some examples of other uses are to limit the input to data received during a mouse-tracking session by setting the mode to NSEventTrackingRunLoopMode, or limit it to data received from a modal panel with NSModalPanelRunLoopMode.

pub fn setRunLoopModes(&self, run_loop_modes: &NSArray<NSRunLoopMode>)

Setter for runLoopModes.

This is copied when set.

pub fn undo(&self)

Closes the top-level undo group if necessary and invokes undoNestedGroup.

This method also invokes endUndoGrouping if the nesting level is 1. Raises an NSInternalInconsistencyException if more than one undo group is open (that is, if the last group isn’t at the top level). This method posts an NSUndoManagerCheckpointNotification.

pub fn redo(&self)

Performs the operations in the last group on the redo stack, if there are any, recording them on the undo stack as a single group.

Raises an NSInternalInconsistencyException if the method is invoked during an undo operation. This method posts an NSUndoManagerCheckpointNotification and NSUndoManagerWillRedoChangeNotification before it performs the redo operation, and it posts the NSUndoManagerDidRedoChangeNotification after it performs the redo operation.

pub fn undoNestedGroup(&self)

Performs the undo operations in the last undo group (whether top-level or nested), recording the operations on the redo stack as a single group.

Raises an NSInternalInconsistencyException if any undo operations have been registered since the last enableUndoRegistration message. This method posts an NSUndoManagerCheckpointNotification and NSUndoManagerWillUndoChangeNotification before it performs the undo operation, and it posts an NSUndoManagerDidUndoChangeNotification after it performs the undo operation.

pub fn canUndo(&self) -> bool

Whether the receiver has any actions to undo.

The return value does not mean you can safely invoke undo or undoNestedGroup — you may have to close open undo groups first.

pub fn canRedo(&self) -> bool

Whether the receiver has any actions to redo.

Because any undo operation registered clears the redo stack, this method posts an NSUndoManagerCheckpointNotification to allow clients to apply their pending operations before testing the redo stack.

pub fn undoCount(&self) -> NSUInteger

How many times undo can be invoked before there are no more actions left to be undone

pub fn redoCount(&self) -> NSUInteger

How many times redo can be invoked before there are no more actions left to be redone

pub fn isUndoing(&self) -> bool

Whether the receiver is in the process of performing its undo or undoNestedGroup method.

pub fn isRedoing(&self) -> bool

Whether the receiver is in the process of performing its redo method.

pub fn removeAllActions(&self)

Clears the undo and redo stacks and re-enables the receiver.

pub unsafe fn removeAllActionsWithTarget(&self, target: &AnyObject)

Clears the undo and redo stacks of all operations involving the specified target as the recipient of the undo message.

Doesn’t re-enable the receiver if it’s disabled.

Parameter target: The recepient of the undo mesages to be removed.

Safety

target should be of the correct type.

pub unsafe fn registerUndoWithTarget_selector_object( &self, target: &AnyObject, selector: Sel, object: Option<&AnyObject>, )

Registers the selector of the specified target to implement a single undo operation that the target receives.

Parameter target: The target of the undo operation. The undo manager maintains an unowned reference to target to prevent retain cycles.

Parameter selector: The selector for the undo operation.

Parameter object: The argument sent with the selector. The undo manager maintains a strong reference to object

Safety

target should be of the correct type.

selector must be a valid selector.

object should be of the correct type.

pub unsafe fn prepareWithInvocationTarget( &self, target: &AnyObject, ) -> Retained<AnyObject>

Prepares the undo manager for invocation-based undo with the given target as the subject of the next undo operation.

For example, when called as:

[[undoManager prepareWithInvocationTarget:target] setFont:oldFont color:oldColor]

When undo is called, the specified target will be called with

[target setFont:oldFont color:oldColor]

Parameter target: The target of the undo operation. The undo manager maintains a weak reference to target.

Returns: A proxy object that forwards messages to the undo manager for recording as undo actions.

Safety

target should be of the correct type.

pub unsafe fn registerUndoWithTarget_handler( &self, target: &AnyObject, undo_handler: &DynBlock<dyn Fn(NonNull<AnyObject>)>, )

Records a single undo operation for a given target so that when an undo is performed, it executes the specified block.

As with other undo operations, this does not strongly retain target. Care should be taken to avoid introducing retain cycles by other references captured by the block.

Parameter target: The target of the undo operation.

Parameter undoHandler: The block to be executed when an operation is undone. The block takes a single argument, the target of the undo operation.

Safety

target should be of the correct type.

pub fn setActionIsDiscardable(&self, discardable: bool)

Sets whether the next undo or redo action is discardable.

Specifies that the latest undo action may be safely discarded when a document can not be saved for any reason. An example might be an undo action that changes the viewable area of a document. To find out if an undo group contains only discardable actions, look for the NSUndoManagerGroupIsDiscardableKey in the userInfo dictionary of the NSUndoManagerWillCloseUndoGroupNotification.

Parameter discardable: Specifies if the action is discardable. YES if the next undo or redo action can be discarded; NO otherwise.

pub fn undoActionIsDiscardable(&self) -> bool

Whether the next undo action is discardable.

Specifies that the latest undo action may be safely discarded when a document can not be saved for any reason. These are typically actions that don’t affect persistent state. An example might be an undo action that changes the viewable area of a document.

pub fn redoActionIsDiscardable(&self) -> bool

Whether the next redo action is discardable.

Specifies that the latest redo action may be safely discarded when a document can not be saved for any reason. These are typically actions that don’t affect persistent state. An example might be an redo action that changes the viewable area of a document.

pub fn undoActionName(&self) -> Retained<NSString>

The name identifying the undo action.

The undo action name. Returns an empty string if no action name has been assigned or if there is nothing to undo. For example, if the menu title is “Undo Delete,” the string returned is “Delete.”

pub fn redoActionName(&self) -> Retained<NSString>

The name identifying the redo action.

The redo action name. Returns an empty string if no action name has been assigned or if there is nothing to redo. For example, if the menu title is “Redo Delete,” the string returned is “Delete.”

pub fn setActionName(&self, action_name: &NSString)

Sets the name of the action associated with the Undo or Redo command.

If actionName is an empty string, the action name currently associated with the menu command is removed. There is no effect if actionName is nil.

Parameter actionName: The name of the action.

pub fn undoActionUserInfoValueForKey( &self, key: &NSUndoManagerUserInfoKey, ) -> Option<Retained<AnyObject>>

Get a value from the undo action’s user info

Parameter key: Which value should be retrieved

pub fn redoActionUserInfoValueForKey( &self, key: &NSUndoManagerUserInfoKey, ) -> Option<Retained<AnyObject>>

Get a value from the redo action’s user info

Parameter key: Which value should be retrieved

pub unsafe fn setActionUserInfoValue_forKey( &self, info: Option<&AnyObject>, key: &NSUndoManagerUserInfoKey, )

Set user info for the Undo or Redo command.

Parameter info: Value to be saved in the user info

Parameter key: Key at which the object should be saved

Safety

info should be of the correct type.

pub fn undoMenuItemTitle(&self) -> Retained<NSString>

The complete title of the Undo menu command, for example, “Undo Paste.”

Returns “Undo” if no action name has been assigned or nil if there is nothing to undo.

pub fn redoMenuItemTitle(&self) -> Retained<NSString>

The complete title of the Redo menu command, for example, “Redo Paste.”

Returns “Redo” if no action name has been assigned or nil if there is nothing to redo.

pub fn undoMenuTitleForUndoActionName( &self, action_name: &NSString, ) -> Retained<NSString>

Returns the complete, localized title of the Undo menu command for the action identified by the given name.

Override this method if you want to customize the localization behaviour. This method is invoked by undoMenuItemTitle.

Parameter actionName: The name of the undo action.

Returns: The localized title of the undo menu item.

pub fn redoMenuTitleForUndoActionName( &self, action_name: &NSString, ) -> Retained<NSString>

Returns the complete, localized title of the Redo menu command for the action identified by the given name.

Override this method if you want to customize the localization behaviour. This method is invoked by redoMenuItemTitle.

Parameter actionName: The name of the redo action.

Returns: The localized title of the redo menu item.

impl NSUndoManager

Methods declared on superclass NSObject.

pub fn init(this: Allocated<Self>) -> Retained<Self>

pub fn new(mtm: MainThreadMarker) -> Retained<Self>