objc2_foundation

Struct NSUserDefaults

pub struct NSUserDefaults { /* private fields */ }

NSUserDefaults is a hierarchical persistent interprocess (optionally distributed) key-value store, optimized for storing user settings.

Hierarchical: NSUserDefaults has a list of places to look for data called the “search list”. A search list is referred to by an arbitrary string called the “suite identifier” or “domain identifier”. When queried, NSUserDefaults checks each entry of its search list until it finds one that contains the key in question, or has searched the whole list. The list is (note: “current host + current user” preferences are unimplemented on iOS, watchOS, and tvOS, and “any user” preferences are not generally useful for applications on those operating systems):

Managed (“forced”) preferences, set by a configuration profile or via mcx from a network administrator

Commandline arguments

Preferences for the current domain, in the cloud

Preferences for the current domain, the current user, in the current host

Preferences for the current domain, the current user, in any host

Preferences added via -addSuiteNamed:

Preferences global to all apps for the current user, in the current host

Preferences global to all apps for the current user, in any host

Preferences for the current domain, for all users, in the current host

Preferences global to all apps for all users, in the current host

Preferences registered with -registerDefaults:

Persistent: Preferences stored in NSUserDefaults persist across reboots and relaunches of apps unless otherwise specified.

Interprocess: Preferences may be accessible to and modified from multiple processes simultaneously (for example between an application and an extension).

Optionally distributed (Currently only supported in Shared iPad for Students mode): Data stored in user defaults can be made “ubiqitous”, i.e. synchronized between devices via the cloud. Ubiquitous user defaults are automatically propagated to all devices logged into the same iCloud account. When reading defaults (via -*ForKey: methods on NSUserDefaults), ubiquitous defaults are searched before local defaults. All operations on ubiquitous defaults are asynchronous, so registered defaults may be returned in place of ubiquitous defaults if downloading from iCloud hasn’t finished. Ubiquitous defaults are specified in the Defaults Configuration File for an application.

Key-Value Store: NSUserDefaults stores Property List objects (NSString, NSData, NSNumber, NSDate, NSArray, and NSDictionary) identified by NSString keys, similar to an NSMutableDictionary.

Optimized for storing user settings: NSUserDefaults is intended for relatively small amounts of data, queried very frequently, and modified occasionally. Using it in other ways may be slow or use more memory than solutions more suited to those uses.

The ‘App’ CFPreferences functions in CoreFoundation act on the same search lists that NSUserDefaults does.

NSUserDefaults can be observed using Key-Value Observing for any key stored in it. Using NSKeyValueObservingOptionPrior to observe changes from other processes or devices will behave as though NSKeyValueObservingOptionPrior was not specified.

See also Apple’s documentation

Implementations

impl NSUserDefaults

pub fn standardUserDefaults() -> Retained<NSUserDefaults>

+standardUserDefaults returns a global instance of NSUserDefaults configured to search the current application’s search list.

pub fn resetStandardUserDefaults()

+resetStandardUserDefaults releases the standardUserDefaults and sets it to nil. A new standardUserDefaults will be created the next time it’s accessed. The only visible effect this has is that all KVO observers of the previous standardUserDefaults will no longer be observing it.

pub fn init(this: Allocated<Self>) -> Retained<Self>

-init is equivalent to -initWithSuiteName:nil

pub fn initWithSuiteName( this: Allocated<Self>, suitename: Option<&NSString>, ) -> Option<Retained<Self>>

-initWithSuiteName: initializes an instance of NSUserDefaults that searches the shared preferences search list for the domain ‘suitename’. For example, using the identifier of an application group will cause the receiver to search the preferences for that group. Passing the current application’s bundle identifier, NSGlobalDomain, or the corresponding CFPreferences constants is an error. Passing nil will search the default search list.

pub fn initWithUser( this: Allocated<Self>, username: &NSString, ) -> Option<Retained<Self>>

👎Deprecated: Use -init instead

-initWithUser: is equivalent to -init

pub fn objectForKey( &self, default_name: &NSString, ) -> Option<Retained<AnyObject>>

-objectForKey: will search the receiver’s search list for a default with the key ‘defaultName’ and return it. If another process has changed defaults in the search list, NSUserDefaults will automatically update to the latest values. If the key in question has been marked as ubiquitous via a Defaults Configuration File, the latest value may not be immediately available, and the registered value will be returned instead.

pub unsafe fn setObject_forKey( &self, value: Option<&AnyObject>, default_name: &NSString, )

-setObject:forKey: immediately stores a value (or removes the value if nil is passed as the value) for the provided key in the search list entry for the receiver’s suite name in the current user and any host, then asynchronously stores the value persistently, where it is made available to other processes.

Safety

value should be of the correct type.

pub fn removeObjectForKey(&self, default_name: &NSString)

-removeObjectForKey: is equivalent to -[… setObject:nil forKey:defaultName]

pub fn stringForKey( &self, default_name: &NSString, ) -> Option<Retained<NSString>>

-stringForKey: is equivalent to -objectForKey:, except that it will convert NSNumber values to their NSString representation. If a non-string non-number value is found, nil will be returned.

pub fn arrayForKey(&self, default_name: &NSString) -> Option<Retained<NSArray>>

-arrayForKey: is equivalent to -objectForKey:, except that it will return nil if the value is not an NSArray.

pub fn dictionaryForKey( &self, default_name: &NSString, ) -> Option<Retained<NSDictionary<NSString, AnyObject>>>

-dictionaryForKey: is equivalent to -objectForKey:, except that it will return nil if the value is not an NSDictionary.

pub fn dataForKey(&self, default_name: &NSString) -> Option<Retained<NSData>>

-dataForKey: is equivalent to -objectForKey:, except that it will return nil if the value is not an NSData.

pub fn stringArrayForKey( &self, default_name: &NSString, ) -> Option<Retained<NSArray<NSString>>>

-stringForKey: is equivalent to -objectForKey:, except that it will return nil if the value is not an NSArray <NSString *>. Note that unlike -stringForKey:, NSNumbers are not converted to NSStrings.

pub fn integerForKey(&self, default_name: &NSString) -> NSInteger

-integerForKey: is equivalent to -objectForKey:, except that it converts the returned value to an NSInteger. If the value is an NSNumber, the result of -integerValue will be returned. If the value is an NSString, it will be converted to NSInteger if possible. If the value is a boolean, it will be converted to either 1 for YES or 0 for NO. If the value is absent or can’t be converted to an integer, 0 will be returned.

pub fn floatForKey(&self, default_name: &NSString) -> c_float

-floatForKey: is similar to -integerForKey:, except that it returns a float, and boolean values will not be converted.

pub fn doubleForKey(&self, default_name: &NSString) -> c_double

-doubleForKey: is similar to -integerForKey:, except that it returns a double, and boolean values will not be converted.

pub fn boolForKey(&self, default_name: &NSString) -> bool

-boolForKey: is equivalent to -objectForKey:, except that it converts the returned value to a BOOL. If the value is an NSNumber, NO will be returned if the value is 0, YES otherwise. If the value is an NSString, values of “YES” or “1” will return YES, and values of “NO”, “0”, or any other string will return NO. If the value is absent or can’t be converted to a BOOL, NO will be returned.

pub fn URLForKey(&self, default_name: &NSString) -> Option<Retained<NSURL>>

-URLForKey: is equivalent to -objectForKey: except that it converts the returned value to an NSURL. If the value is an NSString path, then it will construct a file URL to that path. If the value is an archived URL from -setURL:forKey: it will be unarchived. If the value is absent or can’t be converted to an NSURL, nil will be returned.

pub fn setInteger_forKey(&self, value: NSInteger, default_name: &NSString)

-setInteger:forKey: is equivalent to -setObject:forKey: except that the value is converted from an NSInteger to an NSNumber.

pub fn setFloat_forKey(&self, value: c_float, default_name: &NSString)

-setFloat:forKey: is equivalent to -setObject:forKey: except that the value is converted from a float to an NSNumber.

pub fn setDouble_forKey(&self, value: c_double, default_name: &NSString)

-setDouble:forKey: is equivalent to -setObject:forKey: except that the value is converted from a double to an NSNumber.

pub fn setBool_forKey(&self, value: bool, default_name: &NSString)

-setBool:forKey: is equivalent to -setObject:forKey: except that the value is converted from a BOOL to an NSNumber.

pub fn setURL_forKey(&self, url: Option<&NSURL>, default_name: &NSString)

-setURL:forKey is equivalent to -setObject:forKey: except that the value is archived to an NSData. Use -URLForKey: to retrieve values set this way.

pub unsafe fn registerDefaults( &self, registration_dictionary: &NSDictionary<NSString, AnyObject>, )

-registerDefaults: adds the registrationDictionary to the last item in every search list. This means that after NSUserDefaults has looked for a value in every other valid location, it will look in registered defaults, making them useful as a “fallback” value. Registered defaults are never stored between runs of an application, and are visible only to the application that registers them.

Default values from Defaults Configuration Files will automatically be registered.

Safety

registration_dictionary generic should be of the correct type.

pub fn addSuiteNamed(&self, suite_name: &NSString)

-addSuiteNamed: adds the full search list for ‘suiteName’ as a sub-search-list of the receiver’s. The additional search lists are searched after the current domain, but before global defaults. Passing NSGlobalDomain or the current application’s bundle identifier is unsupported.

pub fn removeSuiteNamed(&self, suite_name: &NSString)

-removeSuiteNamed: removes a sub-searchlist added via -addSuiteNamed:.

pub fn dictionaryRepresentation( &self, ) -> Retained<NSDictionary<NSString, AnyObject>>

-dictionaryRepresentation returns a composite snapshot of the values in the receiver’s search list, such that [[receiver dictionaryRepresentation] objectForKey:x] will return the same thing as [receiver objectForKey:x].

pub fn volatileDomainNames(&self) -> Retained<NSArray<NSString>>

pub fn volatileDomainForName( &self, domain_name: &NSString, ) -> Retained<NSDictionary<NSString, AnyObject>>

pub unsafe fn setVolatileDomain_forName( &self, domain: &NSDictionary<NSString, AnyObject>, domain_name: &NSString, )

Safety

domain generic should be of the correct type.

pub fn removeVolatileDomainForName(&self, domain_name: &NSString)

pub fn persistentDomainNames(&self) -> Retained<NSArray>

👎Deprecated: Not recommended

-persistentDomainNames returns an incomplete list of domains that have preferences stored in them.

pub fn persistentDomainForName( &self, domain_name: &NSString, ) -> Option<Retained<NSDictionary<NSString, AnyObject>>>

-persistentDomainForName: returns a dictionary representation of the search list entry specified by ‘domainName’, the current user, and any host.

pub unsafe fn setPersistentDomain_forName( &self, domain: &NSDictionary<NSString, AnyObject>, domain_name: &NSString, )

-setPersistentDomain:forName: replaces all values in the search list entry specified by ‘domainName’, the current user, and any host, with the values in ‘domain’. The change will be persisted.

Safety

domain generic should be of the correct type.

pub fn removePersistentDomainForName(&self, domain_name: &NSString)

-removePersistentDomainForName: removes all values from the search list entry specified by ‘domainName’, the current user, and any host. The change is persistent.

pub fn synchronize(&self) -> bool

-synchronize is deprecated and will be marked with the API_DEPRECATED macro in a future release.

-synchronize blocks the calling thread until all in-progress set operations have completed. This is no longer necessary. Replacements for previous uses of -synchronize depend on what the intent of calling synchronize was. If you synchronized…

…before reading in order to fetch updated values: remove the synchronize call

…after writing in order to notify another program to read: the other program can use KVO to observe the default without needing to notify

…before exiting in a non-app (command line tool, agent, or daemon) process: call CFPreferencesAppSynchronize(kCFPreferencesCurrentApplication)

…for any other reason: remove the synchronize call

pub fn objectIsForcedForKey(&self, key: &NSString) -> bool

pub fn objectIsForcedForKey_inDomain( &self, key: &NSString, domain: &NSString, ) -> bool

impl NSUserDefaults

Methods declared on superclass NSObject.

pub fn new() -> Retained<Self>