Struct objc2_foundation::NSValue

#[repr(C)]
pub struct NSValue { /* private fields */ }

Implementations

impl NSValue

NSValueGeometryExtensions

pub unsafe fn valueWithPoint(point: NSPoint) -> Retained<NSValue>

pub unsafe fn valueWithSize(size: NSSize) -> Retained<NSValue>

pub unsafe fn valueWithRect(rect: NSRect) -> Retained<NSValue>

pub unsafe fn valueWithEdgeInsets(insets: NSEdgeInsets) -> Retained<NSValue>

pub unsafe fn pointValue(&self) -> NSPoint

pub unsafe fn sizeValue(&self) -> NSSize

pub unsafe fn rectValue(&self) -> NSRect

pub unsafe fn edgeInsetsValue(&self) -> NSEdgeInsets

impl NSValue

NSValueRangeExtensions

pub unsafe fn valueWithRange(range: NSRange) -> Retained<NSValue>

pub unsafe fn rangeValue(&self) -> NSRange

impl NSValue

pub unsafe fn getValue_size(&self, value: NonNull<c_void>, size: NSUInteger)

pub fn objCType(&self) -> NonNull<c_char>

pub unsafe fn initWithBytes_objCType( this: Allocated<Self>, value: NonNull<c_void>, type: NonNull<c_char>, ) -> Retained<Self>

pub unsafe fn initWithCoder( this: Allocated<Self>, coder: &NSCoder, ) -> Option<Retained<Self>>

impl NSValue

NSValueCreation

pub unsafe fn valueWithBytes_objCType( value: NonNull<c_void>, type: NonNull<c_char>, ) -> Retained<NSValue>

pub unsafe fn value_withObjCType( value: NonNull<c_void>, type: NonNull<c_char>, ) -> Retained<NSValue>

impl NSValue

NSValueExtensionMethods

pub unsafe fn valueWithNonretainedObject( an_object: Option<&AnyObject>, ) -> Retained<NSValue>

pub unsafe fn nonretainedObjectValue(&self) -> Option<Retained<AnyObject>>

pub unsafe fn valueWithPointer(pointer: *mut c_void) -> Retained<NSValue>

pub unsafe fn pointerValue(&self) -> *mut c_void

pub fn isEqualToValue(&self, value: &NSValue) -> bool

impl NSValue

NSDeprecated

pub unsafe fn getValue(&self, value: NonNull<c_void>)

👎Deprecated

impl NSValue

Creation methods.

pub fn new<T: 'static + Copy + Encode>(value: T) -> Retained<Self>

Create a new NSValue containing the given type.

Be careful when using this since you may accidentally pass a reference when you wanted to pass a concrete type instead.

Examples

Create an NSValue containing an i32.

use objc2_foundation::NSValue;

let val = NSValue::new(42i32);

impl NSValue

Getter methods.

pub unsafe fn get<T: 'static + Copy + Encode>(&self) -> T

Retrieve the data contained in the NSValue.

Note that this is broken on GNUStep for some types, see gnustep/libs-base#216.

Safety

The type of T must be what the NSValue actually stores, and any safety invariants that the value has must be upheld.

Note that it may be enough, although is not always, to check whether contains_encoding returns true. For example, NonNull<T> have the same encoding as *const T, but NonNull<T> is clearly not safe to return from this function even if you’ve checked the encoding beforehand.

Examples

Store a pointer in NSValue, and retrieve it again afterwards.

use std::ffi::c_void;
use std::ptr;
use objc2_foundation::NSValue;

let val = NSValue::new::<*const c_void>(ptr::null());
// SAFETY: The value was just created with a pointer
let res = unsafe { val.get::<*const c_void>() };
assert!(res.is_null());

pub fn get_range(&self) -> Option<NSRange>

pub fn get_point(&self) -> Option<NSPoint>

pub fn get_size(&self) -> Option<NSSize>

pub fn get_rect(&self) -> Option<NSRect>

pub fn encoding(&self) -> Option<&str>

pub fn contains_encoding<T: 'static + Copy + Encode>(&self) -> bool