objc2_foundation

Struct NSXMLDocument

pub struct NSXMLDocument { /* private fields */ }

An XML Document

Note: if the application of a method would result in more than one element in the children array, an exception is thrown. Trying to add a document, namespace, attribute, or node with a parent also throws an exception. To add a node with a parent first detach or create a copy of it.

See also Apple’s documentation

Implementations

impl NSXMLDocument

pub fn init(this: Allocated<Self>) -> Retained<Self>

pub fn initWithXMLString_options_error( this: Allocated<Self>, string: &NSString, mask: NSXMLNodeOptions, ) -> Result<Retained<Self>, Retained<NSError>>

Returns a document created from either XML or HTML, if the HTMLTidy option is set. Parse errors are returned in error .

pub fn initWithContentsOfURL_options_error( this: Allocated<Self>, url: &NSURL, mask: NSXMLNodeOptions, ) -> Result<Retained<Self>, Retained<NSError>>

Returns a document created from the contents of an XML or HTML URL. Connection problems such as 404, parse errors are returned in error .

pub fn initWithData_options_error( this: Allocated<Self>, data: &NSData, mask: NSXMLNodeOptions, ) -> Result<Retained<Self>, Retained<NSError>>

Returns a document created from data. Parse errors are returned in error .

pub fn initWithRootElement( this: Allocated<Self>, element: Option<&NSXMLElement>, ) -> Retained<Self>

Returns a document with a single child, the root element.

pub unsafe fn replacementClassForClass(cls: &AnyClass) -> &'static AnyClass

Safety

cls probably has further requirements.

pub fn characterEncoding(&self) -> Option<Retained<NSString>>

Sets the character encoding to an IANA type.

pub fn setCharacterEncoding(&self, character_encoding: Option<&NSString>)

Setter for characterEncoding.

This is copied when set.

pub fn version(&self) -> Option<Retained<NSString>>

Sets the XML version. Should be 1.0 or 1.1.

pub fn setVersion(&self, version: Option<&NSString>)

Setter for version.

This is copied when set.

pub fn isStandalone(&self) -> bool

Set whether this document depends on an external DTD. If this option is set the standalone declaration will appear on output.

pub fn setStandalone(&self, standalone: bool)

Setter for isStandalone.

pub fn documentContentKind(&self) -> NSXMLDocumentContentKind

The kind of document.

pub fn setDocumentContentKind( &self, document_content_kind: NSXMLDocumentContentKind, )

Setter for documentContentKind.

pub fn MIMEType(&self) -> Option<Retained<NSString>>

Set the MIME type, eg text/xml.

pub fn setMIMEType(&self, mime_type: Option<&NSString>)

Setter for MIMEType.

This is copied when set.

pub fn DTD(&self) -> Option<Retained<NSXMLDTD>>

Set the associated DTD. This DTD will be output with the document.

pub fn setDTD(&self, dtd: Option<&NSXMLDTD>)

Setter for DTD.

This is copied when set.

pub fn setRootElement(&self, root: &NSXMLElement)

Set the root element. Removes all other children including comments and processing-instructions.

pub fn rootElement(&self) -> Option<Retained<NSXMLElement>>

The root element.

pub fn insertChild_atIndex(&self, child: &NSXMLNode, index: NSUInteger)

Inserts a child at a particular index.

pub fn insertChildren_atIndex( &self, children: &NSArray<NSXMLNode>, index: NSUInteger, )

Insert several children at a particular index.

pub fn removeChildAtIndex(&self, index: NSUInteger)

Removes a child at a particular index.

pub fn setChildren(&self, children: Option<&NSArray<NSXMLNode>>)

Removes all existing children and replaces them with the new children. Set children to nil to simply remove all children.

pub fn addChild(&self, child: &NSXMLNode)

Adds a child to the end of the existing children.

pub fn replaceChildAtIndex_withNode(&self, index: NSUInteger, node: &NSXMLNode)

Replaces a child at a particular index with another child.

pub fn XMLData(&self) -> Retained<NSData>

Invokes XMLDataWithOptions with NSXMLNodeOptionsNone.

pub fn XMLDataWithOptions(&self, options: NSXMLNodeOptions) -> Retained<NSData>

The representation of this node as it would appear in an XML document, encoded based on characterEncoding.

pub fn objectByApplyingXSLT_arguments_error( &self, xslt: &NSData, arguments: Option<&NSDictionary<NSString, NSString>>, ) -> Result<Retained<AnyObject>, Retained<NSError>>

Applies XSLT with arguments (NSString key/value pairs) to this document, returning a new document.

pub fn objectByApplyingXSLTString_arguments_error( &self, xslt: &NSString, arguments: Option<&NSDictionary<NSString, NSString>>, ) -> Result<Retained<AnyObject>, Retained<NSError>>

Applies XSLT as expressed by a string with arguments (NSString key/value pairs) to this document, returning a new document.

pub fn objectByApplyingXSLTAtURL_arguments_error( &self, xslt_url: &NSURL, argument: Option<&NSDictionary<NSString, NSString>>, ) -> Result<Retained<AnyObject>, Retained<NSError>>

Applies the XSLT at a URL with arguments (NSString key/value pairs) to this document, returning a new document. Error may contain a connection error from the URL.

pub fn validateAndReturnError(&self) -> Result<(), Retained<NSError>>

impl NSXMLDocument

Methods declared on superclass NSXMLNode.

pub fn initWithKind( this: Allocated<Self>, kind: NSXMLNodeKind, ) -> Retained<Self>

Invokes

 initWithKind:options:

with options set to NSXMLNodeOptionsNone

pub fn initWithKind_options( this: Allocated<Self>, kind: NSXMLNodeKind, options: NSXMLNodeOptions, ) -> Retained<Self>

Inits a node with fidelity options as description NSXMLNodeOptions.h

impl NSXMLDocument

Methods declared on superclass NSObject.

pub fn new() -> Retained<Self>