objc2_foundation

Trait CopyingHelper

pub unsafe trait CopyingHelper: Message {
    type Result: Message;
}

A helper type for implementing NSCopying.

NSCopying and NSMutableCopying do not in their signatures describe the result type from the copying operation. This is problematic, as it means that using them ends up falling back to AnyObject, which makes copying much less useful and ergonomic.

To properly describe this, we need an associated type which describes the actual result type from a copy. The associated type can’t be present directly on the protocol traits themselves, however, since we want to use them as e.g. ProtocolObject<dyn NSCopying>, so we introduce this helper trait instead. See MutableCopyingHelper for the mutable variant.

We might be able to get rid of this hack once associated type defaults are stabilized.

Safety

The Result type must be correct.

Required Associated Types

type Result: Message

The immutable counterpart of the type, or Self if the type has no immutable counterpart.

The implementation for NSString has itself (NSString) here, while NSMutableString instead has NSString.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.