objc2_foundation

Trait MutableCopyingHelper

pub unsafe trait MutableCopyingHelper: Message {
    type Result: Message;
}

A helper type for implementing NSMutableCopying.

See CopyingHelper for the immutable variant, and more details in general. These traits are split to allow implementing MutableCopyingHelper only when the mutable class is available.

Safety

The Result type must be correct.

Required Associated Types

type Result: Message

The mutable counterpart of the type, or Self if the type has no mutable counterpart.

The implementation for NSString has NSMutableString here, while NSMutableString has itself (NSMutableString).

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.