objc2_foundation

Trait NSMutableCopying

pub unsafe trait NSMutableCopying {
    // Provided methods
    fn mutableCopy(&self) -> Retained<Self::Result>
       where Self: Sized + Message + MutableCopyingHelper { ... }
    unsafe fn mutableCopyWithZone(
        &self,
        zone: *mut NSZone,
    ) -> Retained<Self::Result>
       where Self: Sized + Message + MutableCopyingHelper { ... }
}

A protocol to provide mutable copies of objects.

Only classes that have an “immutable vs. mutable” distinction should adopt this protocol. Use the MutableCopyingHelper trait to specify the return type after copying.

See Apple’s documentation for details.

Example

Implement NSCopying and NSMutableCopying for a class pair like NSString and NSMutableString.

// Immutable copies return NSString

unsafe impl NSCopying for NSString {}
unsafe impl CopyingHelper for NSString {
    type Result = NSString;
}
unsafe impl NSCopying for NSMutableString {}
unsafe impl CopyingHelper for NSMutableString {
    type Result = NSString;
}

// Mutable copies return NSMutableString

unsafe impl NSMutableCopying for NSString {}
unsafe impl MutableCopyingHelper for NSString {
    type Result = NSMutableString;
}
unsafe impl NSMutableCopying for NSMutableString {}
unsafe impl MutableCopyingHelper for NSMutableString {
    type Result = NSMutableString;
}

Provided Methods

fn mutableCopy(&self) -> Retained<Self::Result>
where Self: Sized + Message + MutableCopyingHelper,

Returns a new instance that’s a mutable copy of the receiver.

The output type is the mutable counterpart of the object. E.g. both NSString and NSMutableString return NSMutableString.

unsafe fn mutableCopyWithZone( &self, zone: *mut NSZone, ) -> Retained<Self::Result>
where Self: Sized + Message + MutableCopyingHelper,

Returns a new instance that’s a mutable copy of the receiver.

This is only used when implementing NSMutableCopying, call mutableCopy instead.

Safety

The zone pointer must be valid or NULL.