objc2_foundation

Trait NSObjectNSKeyValueCoding

pub unsafe trait NSObjectNSKeyValueCoding:
    ClassType
    + Sized
    + Sealed {

    // Provided methods
    fn accessInstanceVariablesDirectly() -> bool { ... }
    fn valueForKey(&self, key: &NSString) -> Option<Retained<AnyObject>> { ... }
    unsafe fn setValue_forKey(&self, value: Option<&AnyObject>, key: &NSString) { ... }
    unsafe fn validateValue_forKey_error(
        &self,
        io_value: &mut Option<Retained<AnyObject>>,
        in_key: &NSString,
    ) -> Result<(), Retained<NSError>> { ... }
    unsafe fn mutableArrayValueForKey(
        &self,
        key: &NSString,
    ) -> Retained<NSMutableArray> { ... }
    unsafe fn mutableOrderedSetValueForKey(
        &self,
        key: &NSString,
    ) -> Retained<NSMutableOrderedSet> { ... }
    unsafe fn mutableSetValueForKey(
        &self,
        key: &NSString,
    ) -> Retained<NSMutableSet> { ... }
    fn valueForKeyPath(
        &self,
        key_path: &NSString,
    ) -> Option<Retained<AnyObject>> { ... }
    unsafe fn setValue_forKeyPath(
        &self,
        value: Option<&AnyObject>,
        key_path: &NSString,
    ) { ... }
    unsafe fn validateValue_forKeyPath_error(
        &self,
        io_value: &mut Option<Retained<AnyObject>>,
        in_key_path: &NSString,
    ) -> Result<(), Retained<NSError>> { ... }
    unsafe fn mutableArrayValueForKeyPath(
        &self,
        key_path: &NSString,
    ) -> Retained<NSMutableArray> { ... }
    unsafe fn mutableOrderedSetValueForKeyPath(
        &self,
        key_path: &NSString,
    ) -> Retained<NSMutableOrderedSet> { ... }
    unsafe fn mutableSetValueForKeyPath(
        &self,
        key_path: &NSString,
    ) -> Retained<NSMutableSet> { ... }
    fn valueForUndefinedKey(
        &self,
        key: &NSString,
    ) -> Option<Retained<AnyObject>> { ... }
    unsafe fn setValue_forUndefinedKey(
        &self,
        value: Option<&AnyObject>,
        key: &NSString,
    ) { ... }
    unsafe fn setNilValueForKey(&self, key: &NSString) { ... }
    fn dictionaryWithValuesForKeys(
        &self,
        keys: &NSArray<NSString>,
    ) -> Retained<NSDictionary<NSString, AnyObject>> { ... }
    unsafe fn setValuesForKeysWithDictionary(
        &self,
        keyed_values: &NSDictionary<NSString, AnyObject>,
    ) { ... }

}

Category “NSKeyValueCoding” on NSObject.

Provided Methods

fn accessInstanceVariablesDirectly() -> bool

fn valueForKey(&self, key: &NSString) -> Option<Retained<AnyObject>>

unsafe fn setValue_forKey(&self, value: Option<&AnyObject>, key: &NSString)

Safety

value should be of the correct type.

unsafe fn validateValue_forKey_error( &self, io_value: &mut Option<Retained<AnyObject>>, in_key: &NSString, ) -> Result<(), Retained<NSError>>

Safety

io_value should be of the correct type.

unsafe fn mutableArrayValueForKey( &self, key: &NSString, ) -> Retained<NSMutableArray>

Safety

The returned generic should be of the correct type.

unsafe fn mutableOrderedSetValueForKey( &self, key: &NSString, ) -> Retained<NSMutableOrderedSet>

Safety

The returned generic should be of the correct type.

unsafe fn mutableSetValueForKey(&self, key: &NSString) -> Retained<NSMutableSet>

Safety

The returned generic should be of the correct type.

fn valueForKeyPath(&self, key_path: &NSString) -> Option<Retained<AnyObject>>

unsafe fn setValue_forKeyPath( &self, value: Option<&AnyObject>, key_path: &NSString, )

Safety

value should be of the correct type.

unsafe fn validateValue_forKeyPath_error( &self, io_value: &mut Option<Retained<AnyObject>>, in_key_path: &NSString, ) -> Result<(), Retained<NSError>>

Safety

io_value should be of the correct type.

unsafe fn mutableArrayValueForKeyPath( &self, key_path: &NSString, ) -> Retained<NSMutableArray>

Safety

The returned generic should be of the correct type.

unsafe fn mutableOrderedSetValueForKeyPath( &self, key_path: &NSString, ) -> Retained<NSMutableOrderedSet>

Safety

The returned generic should be of the correct type.

unsafe fn mutableSetValueForKeyPath( &self, key_path: &NSString, ) -> Retained<NSMutableSet>

Safety

The returned generic should be of the correct type.

fn valueForUndefinedKey(&self, key: &NSString) -> Option<Retained<AnyObject>>

unsafe fn setValue_forUndefinedKey( &self, value: Option<&AnyObject>, key: &NSString, )

Safety

value should be of the correct type.

unsafe fn setNilValueForKey(&self, key: &NSString)

fn dictionaryWithValuesForKeys( &self, keys: &NSArray<NSString>, ) -> Retained<NSDictionary<NSString, AnyObject>>

unsafe fn setValuesForKeysWithDictionary( &self, keyed_values: &NSDictionary<NSString, AnyObject>, )

Safety

keyed_values generic should be of the correct type.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.