objc2_foundation

Type Alias NSSize

pub type NSSize = CGSize;

A two-dimensional size.

This is a convenience alias for CGSize. For ease of use, it is available on all platforms, though in practice it is only useful on macOS.

See Apple’s documentation.

Aliased Type

struct NSSize {
    pub width: f64,
    pub height: f64,
}

Fields
width: f64

The dimensions along the x-axis.

height: f64

The dimensions along the y-axis.