Crate objc2_foundation

Bindings to the Foundation framework

See Apple’s docs and the general docs on framework crates for more information.

This is the std equivalent for Objective-C, containing essential data types, collections, and operating-system services.

Rust vs. Objective-C types

A quick overview of some types you will encounter often in Objective-C, and their approximate Rust equivalent.

Objective-C	(approximately) equivalent Rust

NSData*	Arc<[u8]>

NSMutableData*	Vec<u8>

NSString*	Arc<str>

NSMutableString*	String

NSValue*	Arc<dyn Any>

NSNumber*	Arc<enum { I8(i8), U8(u8), I16(i16), U16(u16), I32(i32), U32(u32), I64(i64), U64(u64), F32(f32), F64(f64), CLong(ffi::c_long), CULong(ffi::c_ulong) }>

NSError*	Arc<dyn Error + Send + Sync>

NSException*	Arc<dyn Error + Send + Sync>

NSRange	ops::Range<usize>

NSComparisonResult	cmp::Ordering

NSArray<T>*	Arc<[T]>

NSMutableArray<T>*	Vec<T>

NSDictionary<K, V>*	Arc<HashMap<K, V>>

NSMutableDictionary<K, V>*	HashMap<K, V>

NSEnumerator<T>*	Box<dyn Iterator<T>>

NSCopying*	Box<dyn Clone>

Examples

Basic usage of a few Foundation types.

$ cargo add objc2-foundation --features=all

use objc2_foundation::{ns_string, NSCopying, NSArray};

let string = ns_string!("world");
println!("hello {string}");

let array = NSArray::from_id_slice(&[string.copy()]);
println!("{array:?}");

use objc2::rc::autoreleasepool;
use objc2_foundation::{ns_string, NSArray, NSDictionary, NSObject};

fn main() {
    // Create and compare NSObjects
    let obj = NSObject::new();
    #[allow(clippy::eq_op)]
    {
        println!("{obj:?} == {obj:?}? {:?}", obj == obj);
    }

    let obj2 = NSObject::new();
    println!("{obj:?} == {obj2:?}? {:?}", obj == obj2);

    // Create an NSArray from a Vec
    let objs = vec![obj, obj2];
    let array = NSArray::from_vec(objs);
    for obj in array.iter() {
        println!("{obj:?}");
    }
    println!("{}", array.len());

    // Turn the NSArray back into a Vec
    let mut objs = array.to_vec_retained();
    let obj = objs.pop().unwrap();

    // Create a static NSString
    let string = ns_string!("Hello, world!");
    // Use an autoreleasepool to get the `str` contents of the NSString
    autoreleasepool(|pool| {
        println!("{}", string.as_str(pool));
    });
    // Or use the `Display` implementation
    let _s = string.to_string(); // Using ToString
    println!("{string}"); // Or Display directly

    // Create a dictionary mapping strings to objects
    let keys = &[string];
    let objects = &[obj];
    let dict = NSDictionary::from_id_slice(keys, objects);
    println!("{:?}", dict.get(string));
    println!("{}", dict.len());
}

An example showing how to define your own interfaces to parts that may be missing in the autogenerated interface.

//! Speak synthethized text.
//!
//! This uses `NSSpeechSynthesizer` on macOS, and `AVSpeechSynthesizer` on
//! other Apple platforms. Note that `AVSpeechSynthesizer` _is_ available on
//! macOS, but only since 10.15!
//!
//! Works on macOS >= 10.7 and iOS > 7.0.
#![deny(unsafe_op_in_unsafe_fn)]
#![allow(unused_imports)]

use std::thread;
use std::time::Duration;

use objc2::mutability::InteriorMutable;
use objc2::rc::Retained;
use objc2::{extern_class, msg_send, msg_send_id, ClassType};
use objc2_foundation::{ns_string, NSObject, NSString};

#[cfg(target_os = "macos")]
mod implementation {
    use objc2_foundation::NSCopying;
    use std::cell::Cell;

    use super::*;

    #[link(name = "AppKit", kind = "framework")]
    extern "C" {}

    extern_class!(
        /// <https://developer.apple.com/documentation/appkit/nsspeechsynthesizer?language=objc>
        pub(crate) struct Synthesizer;

        unsafe impl ClassType for Synthesizer {
            type Super = NSObject;
            type Mutability = InteriorMutable;
            const NAME: &'static str = "NSSpeechSynthesizer";
        }
    );

    impl Synthesizer {
        // Uses default voice
        pub(crate) fn new() -> Retained<Self> {
            unsafe { msg_send_id![Self::class(), new] }
        }

        fn set_rate(&self, rate: f32) {
            unsafe { msg_send![self, setRate: rate] }
        }

        fn set_volume(&self, volume: f32) {
            unsafe { msg_send![self, setVolume: volume] }
        }

        fn start_speaking(&self, s: &NSString) {
            let _: bool = unsafe { msg_send![self, startSpeakingString: s] };
        }

        pub(crate) fn speak(&self, utterance: &Utterance) {
            // Convert to the range 90-720 that `NSSpeechSynthesizer` seems to
            // support
            //
            // Note that you'd probably want a nonlinear conversion here to
            // make it match `AVSpeechSynthesizer`.
            self.set_rate(90.0 + (utterance.rate.get() * (360.0 - 90.0)));
            self.set_volume(utterance.volume.get());
            self.start_speaking(&utterance.string);
        }

        pub(crate) fn is_speaking(&self) -> bool {
            unsafe { msg_send![self, isSpeaking] }
        }
    }

    // Shim to make NSSpeechSynthesizer work similar to AVSpeechSynthesizer
    pub(crate) struct Utterance {
        rate: Cell<f32>,
        volume: Cell<f32>,
        string: Retained<NSString>,
    }

    impl Utterance {
        pub(crate) fn new(string: &NSString) -> Self {
            Self {
                rate: Cell::new(0.5),
                volume: Cell::new(1.0),
                string: string.copy(),
            }
        }

        pub(crate) fn set_rate(&self, rate: f32) {
            self.rate.set(rate);
        }

        pub(crate) fn set_volume(&self, volume: f32) {
            self.volume.set(volume);
        }
    }
}

#[cfg(all(target_vendor = "apple", not(target_os = "macos")))]
mod implementation {
    use super::*;

    #[link(name = "AVFoundation", kind = "framework")]
    extern "C" {}

    extern_class!(
        /// <https://developer.apple.com/documentation/avfaudio/avspeechsynthesizer?language=objc>
        #[derive(Debug)]
        pub(crate) struct Synthesizer;

        unsafe impl ClassType for Synthesizer {
            type Super = NSObject;
            type Mutability = InteriorMutable;
            const NAME: &'static str = "AVSpeechSynthesizer";
        }
    );

    impl Synthesizer {
        pub(crate) fn new() -> Retained<Self> {
            unsafe { msg_send_id![Self::class(), new] }
        }

        pub(crate) fn speak(&self, utterance: &Utterance) {
            unsafe { msg_send![self, speakUtterance: utterance] }
        }

        pub(crate) fn is_speaking(&self) -> bool {
            unsafe { msg_send![self, isSpeaking] }
        }
    }

    extern_class!(
        /// <https://developer.apple.com/documentation/avfaudio/avspeechutterance?language=objc>
        #[derive(Debug)]
        pub struct Utterance;

        unsafe impl ClassType for Utterance {
            type Super = NSObject;
            type Mutability = InteriorMutable;
            const NAME: &'static str = "AVSpeechUtterance";
        }
    );

    impl Utterance {
        pub(crate) fn new(string: &NSString) -> Retained<Self> {
            unsafe { msg_send_id![Self::alloc(), initWithString: string] }
        }

        pub(crate) fn set_rate(&self, rate: f32) {
            unsafe { msg_send![self, setRate: rate] }
        }

        pub(crate) fn set_volume(&self, volume: f32) {
            unsafe { msg_send![self, setVolume: volume] }
        }
    }
}

#[cfg(target_vendor = "apple")]
use implementation::{Synthesizer, Utterance};

#[cfg(target_vendor = "apple")]
fn main() {
    let synthesizer = Synthesizer::new();
    let utterance = Utterance::new(ns_string!("Hello from Rust!"));
    utterance.set_rate(0.5);
    utterance.set_volume(0.5);
    synthesizer.speak(&utterance);

    // Wait until speech has properly started up
    thread::sleep(Duration::from_millis(1000));
    // Wait until finished speaking
    while synthesizer.is_speaking() {
        thread::sleep(Duration::from_millis(100));
    }
}

#[cfg(not(target_vendor = "apple"))]
fn main() {
    panic!("this example is only supported on Apple platforms");
}

Modules

array

Utilities for the NSArray and NSMutableArray classes.

dictionary

Utilities for the NSDictionary and NSMutableDictionary classes.

enumerator

Utilities for the NSEnumerator class.

set

Utilities for the NSSet and NSMutableSet classes.

Macros

ns_string

Create a NSString from a static str.

Structs

CGPoint

A point in a two-dimensional coordinate system.

CGRect

The location and dimensions of a rectangle.

CGSize

A two-dimensional size.

MainThreadBound

Make a type that can only be used on the main thread be Send + Sync.

MainThreadMarker

A marker type taken by functions that can only be executed on the main thread.

NSActivityOptions

NSAlignmentOptions

NSArray

NSAttributedString

NSAttributedStringEnumerationOptions

NSAttributedStringFormattingOptions

NSAttributedStringMarkdownInterpretedSyntax

NSAttributedStringMarkdownParsingFailurePolicy

NSAttributedStringMarkdownParsingOptions

NSAttributedStringMarkdownSourcePosition

NSBinarySearchingOptions

NSBlockOperation

NSBundle

NSBundleResourceRequest

NSCoder

NSConstantString

NSCountedSet

NSData

NSDataBase64DecodingOptions

NSDataBase64EncodingOptions

NSDataCompressionAlgorithm

NSDataReadingOptions

NSDataSearchOptions

NSDataWritingOptions

NSDate

NSDecodingFailurePolicy

NSDictionary

NSEdgeInsets

NSEnumerationOptions

NSEnumerator

NSError

NSExtensionContext

NSFastEnumerationState

NSFileSecurity

NSInlinePresentationIntent

NSInvocationOperation

NSMutableArray

NSMutableAttributedString

NSMutableData

NSMutableDictionary

NSMutableSet

NSMutableString

NSNotification

NSNotificationCenter

NSNumber

NSObject

The root class of most Objective-C class hierarchies.

NSOperatingSystemVersion

NSOperation

NSOperationQueue

NSOperationQueuePriority

NSPresentationIntent

NSPresentationIntentKind

NSPresentationIntentTableColumnAlignment

NSProcessInfo

NSProcessInfoThermalState

NSPurgeableData

NSQualityOfService

NSRange

TODO.

NSRectEdge

NSSet

NSSimpleCString

NSSortOptions

NSString

NSStringCompareOptions

NSStringEncodingConversionOptions

NSStringEnumerationOptions

NSThread

NSURL

NSURLBookmarkCreationOptions

NSURLBookmarkResolutionOptions

NSURLComponents

NSURLQueryItem

NSUUID

NSUndoManager

NSUserActivity

NSValue

Enums

NSComparisonResult

Constants that indicate sort order.

Constants

NSASCIIStringEncoding

NSBundleExecutableArchitectureARM64

NSBundleExecutableArchitectureI386

NSBundleExecutableArchitecturePPC

NSBundleExecutableArchitecturePPC64

NSBundleExecutableArchitectureX86_64

NSHPUXOperatingSystemDeprecated

NSISO2022JPStringEncoding

NSISOLatin1StringEncoding

NSISOLatin2StringEncoding

NSJapaneseEUCStringEncoding

NSMACHOperatingSystemDeprecated

NSMacOSRomanStringEncoding

NSNEXTSTEPStringEncoding

NSNonLossyASCIIStringEncoding

NSOSF1OperatingSystemDeprecated

NSShiftJISStringEncoding

NSSolarisOperatingSystemDeprecated

NSSunOSOperatingSystemDeprecated

NSSymbolStringEncoding

NSUTF8StringEncoding

NSUTF16BigEndianStringEncoding

NSUTF16LittleEndianStringEncoding

NSUTF16StringEncoding

NSUTF32BigEndianStringEncoding

NSUTF32LittleEndianStringEncoding

NSUTF32StringEncoding

NSUnicodeStringEncoding

NSWindows95OperatingSystemDeprecated

NSWindowsCP1250StringEncoding

NSWindowsCP1251StringEncoding

NSWindowsCP1252StringEncoding

NSWindowsCP1253StringEncoding

NSWindowsCP1254StringEncoding

NSWindowsNTOperatingSystemDeprecated

Statics

NSAlternateDescriptionAttributeName⚠

NSBundleDidLoadNotification⚠

NSBundleResourceRequestLoadingPriorityUrgent⚠

NSBundleResourceRequestLowDiskSpaceNotification⚠

NSCharacterConversionException⚠

NSCocoaErrorDomain⚠

NSDebugDescriptionErrorKey⚠

NSDidBecomeSingleThreadedNotification⚠

NSEdgeInsetsZero⚠

NSExtensionHostDidBecomeActiveNotification⚠

NSExtensionHostDidEnterBackgroundNotification⚠

NSExtensionHostWillEnterForegroundNotification⚠

NSExtensionHostWillResignActiveNotification⚠

NSExtensionItemsAndErrorsKey⚠

NSFilePathErrorKey⚠

NSFoundationVersionNumber⚠

NSHelpAnchorErrorKey⚠

NSImageURLAttributeName⚠

NSInflectionAgreementArgumentAttributeName⚠

NSInflectionAgreementConceptAttributeName⚠

NSInflectionAlternativeAttributeName⚠

NSInflectionConceptsKey⚠

NSInflectionReferentConceptAttributeName⚠

NSInflectionRuleAttributeName⚠

NSInlinePresentationIntentAttributeName⚠

NSInvocationOperationCancelledException⚠

NSInvocationOperationVoidResultException⚠

NSLanguageIdentifierAttributeName⚠

NSLoadedClasses⚠

NSLocalizedDescriptionKey⚠

NSLocalizedFailureErrorKey⚠

NSLocalizedFailureReasonErrorKey⚠

NSLocalizedRecoveryOptionsErrorKey⚠

NSLocalizedRecoverySuggestionErrorKey⚠

NSMachErrorDomain⚠

NSMarkdownSourcePositionAttributeName⚠

NSMorphologyAttributeName⚠

NSMultipleUnderlyingErrorsKey⚠

NSNotFound

NSOSStatusErrorDomain⚠

NSOperationQueueDefaultMaxConcurrentOperationCount

NSPOSIXErrorDomain⚠

NSParseErrorException⚠

NSPresentationIntentAttributeName⚠

NSProcessInfoPowerStateDidChangeNotification⚠

NSProcessInfoThermalStateDidChangeNotification⚠

NSRecoveryAttempterErrorKey⚠

NSReplacementIndexAttributeName⚠

NSStringEncodingDetectionAllowLossyKey⚠

NSStringEncodingDetectionDisallowedEncodingsKey⚠

NSStringEncodingDetectionFromWindowsKey⚠

NSStringEncodingDetectionLikelyLanguageKey⚠

NSStringEncodingDetectionLossySubstitutionKey⚠

NSStringEncodingDetectionSuggestedEncodingsKey⚠

NSStringEncodingDetectionUseOnlySuggestedEncodingsKey⚠

NSStringEncodingErrorKey⚠

NSStringTransformFullwidthToHalfwidth⚠

NSStringTransformHiraganaToKatakana⚠

NSStringTransformLatinToArabic⚠

NSStringTransformLatinToCyrillic⚠

NSStringTransformLatinToGreek⚠

NSStringTransformLatinToHangul⚠

NSStringTransformLatinToHebrew⚠

NSStringTransformLatinToHiragana⚠

NSStringTransformLatinToKatakana⚠

NSStringTransformLatinToThai⚠

NSStringTransformMandarinToLatin⚠

NSStringTransformStripCombiningMarks⚠

NSStringTransformStripDiacritics⚠

NSStringTransformToLatin⚠

NSStringTransformToUnicodeName⚠

NSStringTransformToXMLHex⚠

NSSystemClockDidChangeNotification⚠

NSThreadWillExitNotification⚠

NSThumbnail1024x1024SizeKey⚠

NSURLAddedToDirectoryDateKey⚠

NSURLApplicationIsScriptableKey⚠

NSURLAttributeModificationDateKey⚠

NSURLCanonicalPathKey⚠

NSURLContentAccessDateKey⚠

NSURLContentModificationDateKey⚠

NSURLContentTypeKey⚠

NSURLCreationDateKey⚠

NSURLCustomIconKey⚠

NSURLDirectoryEntryCountKey⚠

NSURLDocumentIdentifierKey⚠

NSURLEffectiveIconKey⚠

NSURLErrorKey⚠

NSURLFileAllocatedSizeKey⚠

NSURLFileContentIdentifierKey⚠

NSURLFileIdentifierKey⚠

NSURLFileProtectionComplete⚠

NSURLFileProtectionCompleteUnlessOpen⚠

NSURLFileProtectionCompleteUntilFirstUserAuthentication⚠

NSURLFileProtectionCompleteWhenUserInactive⚠

NSURLFileProtectionKey⚠

NSURLFileProtectionNone⚠

NSURLFileResourceIdentifierKey⚠

NSURLFileResourceTypeBlockSpecial⚠

NSURLFileResourceTypeCharacterSpecial⚠

NSURLFileResourceTypeDirectory⚠

NSURLFileResourceTypeKey⚠

NSURLFileResourceTypeNamedPipe⚠

NSURLFileResourceTypeRegular⚠

NSURLFileResourceTypeSocket⚠

NSURLFileResourceTypeSymbolicLink⚠

NSURLFileResourceTypeUnknown⚠

NSURLFileScheme⚠

NSURLFileSecurityKey⚠

NSURLFileSizeKey⚠

NSURLGenerationIdentifierKey⚠

NSURLHasHiddenExtensionKey⚠

NSURLIsAliasFileKey⚠

NSURLIsApplicationKey⚠

NSURLIsDirectoryKey⚠

NSURLIsExcludedFromBackupKey⚠

NSURLIsExecutableKey⚠

NSURLIsHiddenKey⚠

NSURLIsMountTriggerKey⚠

NSURLIsPackageKey⚠

NSURLIsPurgeableKey⚠

NSURLIsReadableKey⚠

NSURLIsRegularFileKey⚠

NSURLIsSparseKey⚠

NSURLIsSymbolicLinkKey⚠

NSURLIsSystemImmutableKey⚠

NSURLIsUbiquitousItemKey⚠

NSURLIsUserImmutableKey⚠

NSURLIsVolumeKey⚠

NSURLIsWritableKey⚠

NSURLKeysOfUnsetValuesKey⚠

NSURLLabelColorKey⚠

NSURLLabelNumberKey⚠

NSURLLinkCountKey⚠

NSURLLocalizedLabelKey⚠

NSURLLocalizedNameKey⚠

NSURLLocalizedTypeDescriptionKey⚠

NSURLMayHaveExtendedAttributesKey⚠

NSURLMayShareFileContentKey⚠

NSURLNameKey⚠

NSURLParentDirectoryURLKey⚠

NSURLPathKey⚠

NSURLPreferredIOBlockSizeKey⚠

NSURLQuarantinePropertiesKey⚠

NSURLTagNamesKey⚠

NSURLThumbnailDictionaryKey⚠

NSURLThumbnailKey⚠

NSURLTotalFileAllocatedSizeKey⚠

NSURLTotalFileSizeKey⚠

NSURLTypeIdentifierKey⚠

NSURLUbiquitousItemContainerDisplayNameKey⚠

NSURLUbiquitousItemDownloadRequestedKey⚠

NSURLUbiquitousItemDownloadingErrorKey⚠

NSURLUbiquitousItemDownloadingStatusCurrent⚠

NSURLUbiquitousItemDownloadingStatusDownloaded⚠

NSURLUbiquitousItemDownloadingStatusKey⚠

NSURLUbiquitousItemDownloadingStatusNotDownloaded⚠

NSURLUbiquitousItemHasUnresolvedConflictsKey⚠

NSURLUbiquitousItemIsDownloadedKey⚠

NSURLUbiquitousItemIsDownloadingKey⚠

NSURLUbiquitousItemIsExcludedFromSyncKey⚠

NSURLUbiquitousItemIsSharedKey⚠

NSURLUbiquitousItemIsUploadedKey⚠

NSURLUbiquitousItemIsUploadingKey⚠

NSURLUbiquitousItemPercentDownloadedKey⚠

NSURLUbiquitousItemPercentUploadedKey⚠

NSURLUbiquitousItemUploadingErrorKey⚠

NSURLUbiquitousSharedItemCurrentUserPermissionsKey⚠

NSURLUbiquitousSharedItemCurrentUserRoleKey⚠

NSURLUbiquitousSharedItemMostRecentEditorNameComponentsKey⚠

NSURLUbiquitousSharedItemOwnerNameComponentsKey⚠

NSURLUbiquitousSharedItemPermissionsReadOnly⚠

NSURLUbiquitousSharedItemPermissionsReadWrite⚠

NSURLUbiquitousSharedItemRoleOwner⚠

NSURLUbiquitousSharedItemRoleParticipant⚠

NSURLVolumeAvailableCapacityForImportantUsageKey⚠

NSURLVolumeAvailableCapacityForOpportunisticUsageKey⚠

NSURLVolumeAvailableCapacityKey⚠

NSURLVolumeCreationDateKey⚠

NSURLVolumeIdentifierKey⚠

NSURLVolumeIsAutomountedKey⚠

NSURLVolumeIsBrowsableKey⚠

NSURLVolumeIsEjectableKey⚠

NSURLVolumeIsEncryptedKey⚠

NSURLVolumeIsInternalKey⚠

NSURLVolumeIsJournalingKey⚠

NSURLVolumeIsLocalKey⚠

NSURLVolumeIsReadOnlyKey⚠

NSURLVolumeIsRemovableKey⚠

NSURLVolumeIsRootFileSystemKey⚠

NSURLVolumeLocalizedFormatDescriptionKey⚠

NSURLVolumeLocalizedNameKey⚠

NSURLVolumeMaximumFileSizeKey⚠

NSURLVolumeMountFromLocationKey⚠

NSURLVolumeNameKey⚠

NSURLVolumeResourceCountKey⚠

NSURLVolumeSubtypeKey⚠

NSURLVolumeSupportsAccessPermissionsKey⚠

NSURLVolumeSupportsAdvisoryFileLockingKey⚠

NSURLVolumeSupportsCasePreservedNamesKey⚠

NSURLVolumeSupportsCaseSensitiveNamesKey⚠

NSURLVolumeSupportsCompressionKey⚠

NSURLVolumeSupportsExclusiveRenamingKey⚠

NSURLVolumeSupportsExtendedSecurityKey⚠

NSURLVolumeSupportsFileCloningKey⚠

NSURLVolumeSupportsFileProtectionKey⚠

NSURLVolumeSupportsHardLinksKey⚠

NSURLVolumeSupportsImmutableFilesKey⚠

NSURLVolumeSupportsJournalingKey⚠

NSURLVolumeSupportsPersistentIDsKey⚠

NSURLVolumeSupportsRenamingKey⚠

NSURLVolumeSupportsRootDirectoryDatesKey⚠

NSURLVolumeSupportsSparseFilesKey⚠

NSURLVolumeSupportsSwapRenamingKey⚠

NSURLVolumeSupportsSymbolicLinksKey⚠

NSURLVolumeSupportsVolumeSizesKey⚠

NSURLVolumeSupportsZeroRunsKey⚠

NSURLVolumeTotalCapacityKey⚠

NSURLVolumeTypeNameKey⚠

NSURLVolumeURLForRemountingKey⚠

NSURLVolumeURLKey⚠

NSURLVolumeUUIDStringKey⚠

NSUnderlyingErrorKey⚠

NSUndoCloseGroupingRunLoopOrdering

NSUndoManagerCheckpointNotification⚠

NSUndoManagerDidCloseUndoGroupNotification⚠

NSUndoManagerDidOpenUndoGroupNotification⚠

NSUndoManagerDidRedoChangeNotification⚠

NSUndoManagerDidUndoChangeNotification⚠

NSUndoManagerGroupIsDiscardableKey⚠

NSUndoManagerWillCloseUndoGroupNotification⚠

NSUndoManagerWillRedoChangeNotification⚠

NSUndoManagerWillUndoChangeNotification⚠

NSUserActivityTypeBrowsingWeb⚠

NSWillBecomeMultiThreadedNotification⚠

NSZeroPoint⚠

NSZeroRect⚠

NSZeroSize⚠

Traits

NSCoding

NSCopying

A protocol to provide functional copies of objects.

NSDiscardableContent

NSExtensionRequestHandling

NSFastEnumeration

NSMutableCopying

A protocol to provide mutable copies of objects.

NSObjectNSCoderMethods

Category “NSCoderMethods” on NSObject.

NSObjectNSDiscardableContentProxy

Category “NSDiscardableContentProxy” on NSObject.

NSObjectNSErrorRecoveryAttempting

Category “NSErrorRecoveryAttempting” on NSObject.

NSObjectNSThreadPerformAdditions

Category “NSThreadPerformAdditions” on NSObject.

NSObjectProtocol

The methods that are fundamental to most Objective-C objects.

NSSecureCoding

NSUserActivityDelegate

Functions

NSClassFromString⚠

NSContainsRect⚠

NSDeallocateObject⚠

NSDecrementExtraRefCountWasZero⚠

NSDivideRect⚠

NSEdgeInsetsEqual⚠

NSEqualPoints⚠

NSEqualRects⚠

NSEqualSizes⚠

NSExtraRefCount⚠

NSGetSizeAndAlignment⚠

NSIncrementExtraRefCount⚠

NSInsetRect⚠

NSIntegralRect⚠

NSIntegralRectWithOptions⚠

NSIntersectionRange⚠

NSIntersectionRect⚠

NSIntersectsRect⚠

NSIsEmptyRect⚠

NSMouseInRect⚠

NSOffsetRect⚠

NSPointFromString⚠

NSPointInRect⚠

NSProtocolFromString⚠

NSRangeFromString⚠

NSRectFromString⚠

NSSelectorFromString⚠

NSSizeFromString⚠

NSStringFromClass⚠

NSStringFromPoint⚠

NSStringFromProtocol⚠

NSStringFromRange⚠

NSStringFromRect⚠

NSStringFromSelector⚠

NSStringFromSize⚠

NSUnionRange⚠

NSUnionRect⚠

NXReadNSObjectFromCoder⚠Deprecated

is_main_thread

Whether the current thread is the main thread.

is_multi_threaded

Whether the application is multithreaded according to Cocoa.

run_on_main

Submit the given closure to the runloop on the main thread.

Type Aliases

CGFloat

The basic type for all floating-point values.

NSAttributedStringFormattingContextKey

NSAttributedStringKey

NSComparator

NSErrorDomain

NSErrorUserInfoKey

NSExceptionName

NSInteger

A signed integer value type.

NSNotificationName

NSPoint

A point in a Cartesian coordinate system.

NSPointArray

NSPointPointer

NSRangePointer

NSRect

A rectangle.

NSRectArray

NSRectPointer

NSRunLoopMode

NSSize

A two-dimensional size.

NSSizeArray

NSSizePointer

NSStringEncoding

NSStringEncodingDetectionOptionsKey

NSStringTransform

NSTimeInterval

NSUInteger

Describes an unsigned integer.

NSURLBookmarkFileCreationOptions

NSURLFileProtectionType

NSURLFileResourceType

NSURLResourceKey

NSURLThumbnailDictionaryItem

NSURLUbiquitousItemDownloadingStatus

NSURLUbiquitousSharedItemPermissions

NSURLUbiquitousSharedItemRole

NSUserActivityPersistentIdentifier

OSType

unichar