Struct objc2_foundation::NSMutableDictionary

#[repr(C)]
pub struct NSMutableDictionary<KeyType: ?Sized = AnyObject, ObjectType: ?Sized = AnyObject> { /* private fields */ }

Implementations

impl<K: Message + Eq + Hash + HasStableHash, V: Message> NSMutableDictionary<K, V>

pub fn from_vec<Q>(keys: &[&Q], objects: Vec<Retained<V>>) -> Retained<Self>
where Q: Message + NSCopying + CounterpartOrSelf<Immutable = K>,

pub fn from_id_slice<Q>(keys: &[&Q], objects: &[Retained<V>]) -> Retained<Self>
where Q: Message + NSCopying + CounterpartOrSelf<Immutable = K>, V: IsIdCloneable,

pub fn from_slice<Q>(keys: &[&Q], objects: &[&V]) -> Retained<Self>
where Q: Message + NSCopying + CounterpartOrSelf<Immutable = K>, V: IsRetainable,

impl<K: Message + Eq + Hash + HasStableHash, V: Message> NSMutableDictionary<K, V>

pub fn insert_id(&mut self, key: &K, value: Retained<V>) -> Option<Retained<V>>
where K: NSCopying + CounterpartOrSelf<Immutable = K>,

Inserts a key-value pair into the dictionary.

If the dictionary did not have this key present, None is returned. If the dictionary did have this key present, the value is updated, and the old value is returned.

Examples

use objc2_foundation::{NSMutableDictionary, NSObject, ns_string};

let mut dict = NSMutableDictionary::new();
dict.insert_id(ns_string!("one"), NSObject::new());

pub fn insert(&mut self, key: &K, value: &V) -> Option<Retained<V>>
where K: NSCopying + CounterpartOrSelf<Immutable = K>, V: IsRetainable,

Inserts a key-value pair into the dictionary.

If the dictionary did not have this key present, None is returned. If the dictionary did have this key present, the value is updated, and the old value is returned.

Examples

use objc2_foundation::{ns_string, NSCopying, NSMutableDictionary};

let mut dict = NSMutableDictionary::new();
dict.insert_id(ns_string!("key"), ns_string!("value").copy());

pub fn remove(&mut self, key: &K) -> Option<Retained<V>>
where K: CounterpartOrSelf<Immutable = K>,

Removes a key from the dictionary, returning the value at the key if the key was previously in the dictionary.

Examples

use objc2_foundation::{ns_string, NSMutableDictionary, NSObject};

let mut dict = NSMutableDictionary::new();
dict.insert_id(ns_string!("one"), NSObject::new());
dict.remove(ns_string!("one"));
assert!(dict.is_empty());

impl<KeyType: Message, ObjectType: Message> NSMutableDictionary<KeyType, ObjectType>

Methods declared on superclass NSDictionary

NSDictionaryCreation

pub unsafe fn dictionary() -> Retained<Self>

pub unsafe fn dictionaryWithObject_forKey( object: &ObjectType, key: &ProtocolObject<dyn NSCopying>, ) -> Retained<Self>

pub unsafe fn dictionaryWithObjects_forKeys_count( objects: *mut NonNull<ObjectType>, keys: *mut NonNull<ProtocolObject<dyn NSCopying>>, cnt: NSUInteger, ) -> Retained<Self>

pub unsafe fn dictionaryWithDictionary( dict: &NSDictionary<KeyType, ObjectType>, ) -> Retained<Self>

pub unsafe fn dictionaryWithObjects_forKeys( objects: &NSArray<ObjectType>, keys: &NSArray<ProtocolObject<dyn NSCopying>>, ) -> Retained<Self>

pub unsafe fn initWithDictionary( this: Allocated<Self>, other_dictionary: &NSDictionary<KeyType, ObjectType>, ) -> Retained<Self>

pub unsafe fn initWithDictionary_copyItems( this: Allocated<Self>, other_dictionary: &NSDictionary<KeyType, ObjectType>, flag: bool, ) -> Retained<Self>

pub unsafe fn initWithObjects_forKeys( this: Allocated<Self>, objects: &NSArray<ObjectType>, keys: &NSArray<ProtocolObject<dyn NSCopying>>, ) -> Retained<Self>

impl<KeyType: Message, ObjectType: Message> NSMutableDictionary<KeyType, ObjectType>

pub fn removeObjectForKey(&mut self, a_key: &KeyType)

pub unsafe fn setObject_forKey( &mut self, an_object: &ObjectType, a_key: &ProtocolObject<dyn NSCopying>, )

pub fn init(this: Allocated<Self>) -> Retained<Self>

pub unsafe fn initWithCapacity( this: Allocated<Self>, num_items: NSUInteger, ) -> Retained<Self>

pub unsafe fn initWithCoder( this: Allocated<Self>, coder: &NSCoder, ) -> Option<Retained<Self>>

impl<KeyType: Message, ObjectType: Message> NSMutableDictionary<KeyType, ObjectType>

Methods declared on superclass NSDictionary

pub unsafe fn initWithObjects_forKeys_count( this: Allocated<Self>, objects: *mut NonNull<ObjectType>, keys: *mut NonNull<ProtocolObject<dyn NSCopying>>, cnt: NSUInteger, ) -> Retained<Self>

impl<KeyType: Message, ObjectType: Message> NSMutableDictionary<KeyType, ObjectType>

Methods declared on superclass NSObject

pub fn new() -> Retained<Self>

impl<KeyType: Message, ObjectType: Message> NSMutableDictionary<KeyType, ObjectType>

NSExtendedMutableDictionary

pub unsafe fn addEntriesFromDictionary( &mut self, other_dictionary: &NSDictionary<KeyType, ObjectType>, )

pub fn removeAllObjects(&mut self)

pub unsafe fn removeObjectsForKeys(&mut self, key_array: &NSArray<KeyType>)

pub unsafe fn setDictionary( &mut self, other_dictionary: &NSDictionary<KeyType, ObjectType>, )

pub unsafe fn setObject_forKeyedSubscript( &mut self, obj: Option<&ObjectType>, key: &ProtocolObject<dyn NSCopying>, )

impl<KeyType: Message, ObjectType: Message> NSMutableDictionary<KeyType, ObjectType>

NSMutableDictionaryCreation

pub unsafe fn dictionaryWithCapacity(num_items: NSUInteger) -> Retained<Self>

pub unsafe fn dictionaryWithContentsOfFile( path: &NSString, ) -> Option<Retained<NSMutableDictionary<KeyType, ObjectType>>>

pub unsafe fn dictionaryWithContentsOfURL( url: &NSURL, ) -> Option<Retained<NSMutableDictionary<KeyType, ObjectType>>>

pub unsafe fn initWithContentsOfFile( this: Allocated<Self>, path: &NSString, ) -> Option<Retained<NSMutableDictionary<KeyType, ObjectType>>>

pub unsafe fn initWithContentsOfURL( this: Allocated<Self>, url: &NSURL, ) -> Option<Retained<NSMutableDictionary<KeyType, ObjectType>>>

impl<KeyType: Message, ObjectType: Message> NSMutableDictionary<KeyType, ObjectType>

NSSharedKeySetDictionary

pub unsafe fn dictionaryWithSharedKeySet( keyset: &AnyObject, ) -> Retained<NSMutableDictionary<KeyType, ObjectType>>

impl<KeyType: Message, ObjectType: Message> NSMutableDictionary<KeyType, ObjectType>

NSKeyValueCoding

pub unsafe fn setValue_forKey( &mut self, value: Option<&ObjectType>, key: &NSString, )