Struct objc2_foundation::NSObject

#[repr(C)]
pub struct NSObject { /* private fields */ }

The root class of most Objective-C class hierarchies.

This represents the NSObject class. The name “NSObject” also refers to a protocol, see NSObjectProtocol for that.

Since this class is only available with the Foundation framework, objc2 links to it for you.

This is exported under objc2_foundation::NSObject, you probably want to use that path instead.

Implementations

impl NSObject

pub fn new() -> Retained<NSObject>

Create a new empty NSObject.

This method is a shorthand for calling alloc and then init.

pub fn init(this: Allocated<NSObject>) -> Retained<NSObject>

Initialize an already allocated object.

See Apple’s documentation for details.

Example

use objc2::runtime::NSObject;
use objc2::ClassType;

let obj = NSObject::init(NSObject::alloc());

pub fn doesNotRecognizeSelector(&self, sel: Sel) -> !

Handle messages the object doesn’t recognize.

See Apple’s documentation for details.