Struct objc2_foundation::NSURL

#[repr(C)]
pub struct NSURL { /* private fields */ }

Implementations

impl NSURL

pub unsafe fn initWithScheme_host_path( this: Allocated<Self>, scheme: &NSString, host: Option<&NSString>, path: &NSString, ) -> Option<Retained<Self>>

👎Deprecated: Use NSURLComponents instead, which lets you create a valid URL with any valid combination of URL components and subcomponents (not just scheme, host and path), and lets you set components and subcomponents with either percent-encoded or un-percent-encoded strings.

pub unsafe fn initFileURLWithPath_isDirectory_relativeToURL( this: Allocated<Self>, path: &NSString, is_dir: bool, base_url: Option<&NSURL>, ) -> Retained<Self>

pub unsafe fn initFileURLWithPath_relativeToURL( this: Allocated<Self>, path: &NSString, base_url: Option<&NSURL>, ) -> Retained<Self>

pub unsafe fn initFileURLWithPath_isDirectory( this: Allocated<Self>, path: &NSString, is_dir: bool, ) -> Retained<Self>

pub unsafe fn initFileURLWithPath( this: Allocated<Self>, path: &NSString, ) -> Retained<Self>

pub unsafe fn fileURLWithPath_isDirectory_relativeToURL( path: &NSString, is_dir: bool, base_url: Option<&NSURL>, ) -> Retained<NSURL>

pub unsafe fn fileURLWithPath_relativeToURL( path: &NSString, base_url: Option<&NSURL>, ) -> Retained<NSURL>

pub unsafe fn fileURLWithPath_isDirectory( path: &NSString, is_dir: bool, ) -> Retained<NSURL>

pub unsafe fn fileURLWithPath(path: &NSString) -> Retained<NSURL>

pub unsafe fn initFileURLWithFileSystemRepresentation_isDirectory_relativeToURL( this: Allocated<Self>, path: NonNull<c_char>, is_dir: bool, base_url: Option<&NSURL>, ) -> Retained<Self>

pub unsafe fn fileURLWithFileSystemRepresentation_isDirectory_relativeToURL( path: NonNull<c_char>, is_dir: bool, base_url: Option<&NSURL>, ) -> Retained<NSURL>

pub unsafe fn initWithString( this: Allocated<Self>, url_string: &NSString, ) -> Option<Retained<Self>>

pub unsafe fn initWithString_relativeToURL( this: Allocated<Self>, url_string: &NSString, base_url: Option<&NSURL>, ) -> Option<Retained<Self>>

pub unsafe fn URLWithString(url_string: &NSString) -> Option<Retained<Self>>

pub unsafe fn URLWithString_relativeToURL( url_string: &NSString, base_url: Option<&NSURL>, ) -> Option<Retained<Self>>

pub unsafe fn initWithString_encodingInvalidCharacters( this: Allocated<Self>, url_string: &NSString, encoding_invalid_characters: bool, ) -> Option<Retained<Self>>

pub unsafe fn URLWithString_encodingInvalidCharacters( url_string: &NSString, encoding_invalid_characters: bool, ) -> Option<Retained<Self>>

pub unsafe fn initWithDataRepresentation_relativeToURL( this: Allocated<Self>, data: &NSData, base_url: Option<&NSURL>, ) -> Retained<Self>

pub unsafe fn URLWithDataRepresentation_relativeToURL( data: &NSData, base_url: Option<&NSURL>, ) -> Retained<NSURL>

pub unsafe fn initAbsoluteURLWithDataRepresentation_relativeToURL( this: Allocated<Self>, data: &NSData, base_url: Option<&NSURL>, ) -> Retained<Self>

pub unsafe fn absoluteURLWithDataRepresentation_relativeToURL( data: &NSData, base_url: Option<&NSURL>, ) -> Retained<NSURL>

pub unsafe fn dataRepresentation(&self) -> Retained<NSData>

pub unsafe fn absoluteString(&self) -> Option<Retained<NSString>>

pub unsafe fn relativeString(&self) -> Retained<NSString>

pub unsafe fn baseURL(&self) -> Option<Retained<NSURL>>

pub unsafe fn absoluteURL(&self) -> Option<Retained<NSURL>>

pub unsafe fn scheme(&self) -> Option<Retained<NSString>>

pub unsafe fn resourceSpecifier(&self) -> Option<Retained<NSString>>

pub unsafe fn host(&self) -> Option<Retained<NSString>>

pub unsafe fn port(&self) -> Option<Retained<NSNumber>>

pub unsafe fn user(&self) -> Option<Retained<NSString>>

pub unsafe fn password(&self) -> Option<Retained<NSString>>

pub unsafe fn path(&self) -> Option<Retained<NSString>>

pub unsafe fn fragment(&self) -> Option<Retained<NSString>>

pub unsafe fn parameterString(&self) -> Option<Retained<NSString>>

👎Deprecated: The parameterString method is deprecated. Post deprecation for applications linked with or after the macOS 10.15, and for all iOS, watchOS, and tvOS applications, parameterString will always return nil, and the path method will return the complete path including the semicolon separator and params component if the URL string contains them.

pub unsafe fn query(&self) -> Option<Retained<NSString>>

pub unsafe fn relativePath(&self) -> Option<Retained<NSString>>

pub unsafe fn hasDirectoryPath(&self) -> bool

pub unsafe fn getFileSystemRepresentation_maxLength( &self, buffer: NonNull<c_char>, max_buffer_length: NSUInteger, ) -> bool

pub unsafe fn fileSystemRepresentation(&self) -> NonNull<c_char>

pub unsafe fn isFileURL(&self) -> bool

pub unsafe fn standardizedURL(&self) -> Option<Retained<NSURL>>

pub unsafe fn isFileReferenceURL(&self) -> bool

pub unsafe fn fileReferenceURL(&self) -> Option<Retained<NSURL>>

pub unsafe fn filePathURL(&self) -> Option<Retained<NSURL>>

pub unsafe fn getResourceValue_forKey_error( &self, value: &mut Option<Retained<AnyObject>>, key: &NSURLResourceKey, ) -> Result<(), Retained<NSError>>

pub unsafe fn resourceValuesForKeys_error( &self, keys: &NSArray<NSURLResourceKey>, ) -> Result<Retained<NSDictionary<NSURLResourceKey, AnyObject>>, Retained<NSError>>

pub unsafe fn setResourceValue_forKey_error( &self, value: Option<&AnyObject>, key: &NSURLResourceKey, ) -> Result<(), Retained<NSError>>

pub unsafe fn setResourceValues_error( &self, keyed_values: &NSDictionary<NSURLResourceKey, AnyObject>, ) -> Result<(), Retained<NSError>>

pub unsafe fn removeCachedResourceValueForKey(&self, key: &NSURLResourceKey)

pub unsafe fn removeAllCachedResourceValues(&self)

pub unsafe fn setTemporaryResourceValue_forKey( &self, value: Option<&AnyObject>, key: &NSURLResourceKey, )

pub unsafe fn bookmarkDataWithOptions_includingResourceValuesForKeys_relativeToURL_error( &self, options: NSURLBookmarkCreationOptions, keys: Option<&NSArray<NSURLResourceKey>>, relative_url: Option<&NSURL>, ) -> Result<Retained<NSData>, Retained<NSError>>

pub unsafe fn initByResolvingBookmarkData_options_relativeToURL_bookmarkDataIsStale_error( this: Allocated<Self>, bookmark_data: &NSData, options: NSURLBookmarkResolutionOptions, relative_url: Option<&NSURL>, is_stale: *mut Bool, ) -> Result<Retained<Self>, Retained<NSError>>

pub unsafe fn URLByResolvingBookmarkData_options_relativeToURL_bookmarkDataIsStale_error( bookmark_data: &NSData, options: NSURLBookmarkResolutionOptions, relative_url: Option<&NSURL>, is_stale: *mut Bool, ) -> Result<Retained<Self>, Retained<NSError>>

pub unsafe fn resourceValuesForKeys_fromBookmarkData( keys: &NSArray<NSURLResourceKey>, bookmark_data: &NSData, ) -> Option<Retained<NSDictionary<NSURLResourceKey, AnyObject>>>

pub unsafe fn writeBookmarkData_toURL_options_error( bookmark_data: &NSData, bookmark_file_url: &NSURL, options: NSURLBookmarkFileCreationOptions, ) -> Result<(), Retained<NSError>>

pub unsafe fn bookmarkDataWithContentsOfURL_error( bookmark_file_url: &NSURL, ) -> Result<Retained<NSData>, Retained<NSError>>

pub unsafe fn URLByResolvingAliasFileAtURL_options_error( url: &NSURL, options: NSURLBookmarkResolutionOptions, ) -> Result<Retained<Self>, Retained<NSError>>

pub unsafe fn startAccessingSecurityScopedResource(&self) -> bool

pub unsafe fn stopAccessingSecurityScopedResource(&self)

impl NSURL

Methods declared on superclass NSObject

pub unsafe fn init(this: Allocated<Self>) -> Retained<Self>

pub unsafe fn new() -> Retained<Self>

impl NSURL

NSPromisedItems

pub unsafe fn getPromisedItemResourceValue_forKey_error( &self, value: &mut Option<Retained<AnyObject>>, key: &NSURLResourceKey, ) -> Result<(), Retained<NSError>>

pub unsafe fn promisedItemResourceValuesForKeys_error( &self, keys: &NSArray<NSURLResourceKey>, ) -> Result<Retained<NSDictionary<NSURLResourceKey, AnyObject>>, Retained<NSError>>

pub unsafe fn checkPromisedItemIsReachableAndReturnError( &self, ) -> Result<(), Retained<NSError>>

impl NSURL

This impl block contains no items.

NSItemProvider

impl NSURL

NSURLPathUtilities

pub unsafe fn fileURLWithPathComponents( components: &NSArray<NSString>, ) -> Option<Retained<NSURL>>

pub unsafe fn pathComponents(&self) -> Option<Retained<NSArray<NSString>>>

pub unsafe fn lastPathComponent(&self) -> Option<Retained<NSString>>

pub unsafe fn pathExtension(&self) -> Option<Retained<NSString>>

pub unsafe fn URLByAppendingPathComponent( &self, path_component: &NSString, ) -> Option<Retained<NSURL>>

pub unsafe fn URLByAppendingPathComponent_isDirectory( &self, path_component: &NSString, is_directory: bool, ) -> Option<Retained<NSURL>>

pub unsafe fn URLByDeletingLastPathComponent(&self) -> Option<Retained<NSURL>>

pub unsafe fn URLByAppendingPathExtension( &self, path_extension: &NSString, ) -> Option<Retained<NSURL>>

pub unsafe fn URLByDeletingPathExtension(&self) -> Option<Retained<NSURL>>

pub unsafe fn checkResourceIsReachableAndReturnError( &self, ) -> Result<(), Retained<NSError>>

pub unsafe fn URLByStandardizingPath(&self) -> Option<Retained<NSURL>>

pub unsafe fn URLByResolvingSymlinksInPath(&self) -> Option<Retained<NSURL>>

impl NSURL

NSURLLoading

pub unsafe fn resourceDataUsingCache( &self, should_use_cache: bool, ) -> Option<Retained<NSData>>

👎Deprecated: Use NSURLConnection instead

pub unsafe fn loadResourceDataNotifyingClient_usingCache( &self, client: &AnyObject, should_use_cache: bool, )

👎Deprecated: Use NSURLConnection instead

pub unsafe fn propertyForKey( &self, property_key: &NSString, ) -> Option<Retained<AnyObject>>

👎Deprecated: Use NSURLConnection instead

pub unsafe fn setResourceData(&self, data: &NSData) -> bool

👎Deprecated: Use NSURLConnection instead

pub unsafe fn setProperty_forKey( &self, property: &AnyObject, property_key: &NSString, ) -> bool

👎Deprecated: Use NSURLConnection instead