objc2_foundation

Trait NSCopying

pub unsafe trait NSCopying {
    // Provided methods
    fn copy(&self) -> Retained<Self::Result>
       where Self: Sized + Message + CopyingHelper { ... }
    unsafe fn copyWithZone(&self, zone: *mut NSZone) -> Retained<Self::Result>
       where Self: Sized + Message + CopyingHelper { ... }
}

A protocol to provide functional copies of objects.

This is similar to Rust’s Clone trait, along with sharing a few similarities to the std::borrow::ToOwned trait with regards to the output type.

To allow using this in a meaningful way in Rust, we have to “enrich” the implementation by also specifying the resulting type, see CopyingHelper for details.

See also Apple’s documentation.

Examples

Implement NSCopying for an externally defined class.

use objc2::extern_class;
use objc2_foundation::{CopyingHelper, NSCopying, NSObject};

extern_class!(
    #[unsafe(super(NSObject))]
    struct ExampleClass;
);

unsafe impl NSCopying for ExampleClass {}

// Copying ExampleClass returns another ExampleClass.
unsafe impl CopyingHelper for ExampleClass {
    type Result = Self;
}

Implement NSCopying for a custom class.

use objc2::{define_class, msg_send, AnyThread, DefinedClass};
use objc2::rc::Retained;
use objc2::runtime::NSZone;
use objc2_foundation::{CopyingHelper, NSCopying, NSObject};

define_class!(
    #[unsafe(super(NSObject))]
    struct CustomClass;

    unsafe impl NSCopying for CustomClass {
        #[unsafe(method_id(copyWithZone:))]
        fn copyWithZone(&self, _zone: *const NSZone) -> Retained<Self> {
            // Create new class, and transfer ivars
            let new = Self::alloc().set_ivars(self.ivars().clone());
            unsafe { msg_send![super(new), init] }
        }
    }
);

// Copying CustomClass returns another CustomClass.
unsafe impl CopyingHelper for CustomClass {
    type Result = Self;
}

Provided Methods

fn copy(&self) -> Retained<Self::Result>
where Self: Sized + Message + CopyingHelper,

Returns a new instance that’s a copy of the receiver.

The output type is the immutable counterpart of the object, which is usually Self, but e.g. NSMutableString returns NSString.

unsafe fn copyWithZone(&self, zone: *mut NSZone) -> Retained<Self::Result>
where Self: Sized + Message + CopyingHelper,

Returns a new instance that’s a copy of the receiver.

This is only used when implementing NSCopying, call copy instead.

Safety

The zone pointer must be valid or NULL.