objc2_foundation

Trait NSCopying

pub unsafe trait NSCopying {
    // Provided method
    fn copy(&self) -> Retained<Self::Immutable>
       where Self: Sized + Message + CounterpartOrSelf { ... }
}

A protocol to provide functional copies of objects.

This is similar to Rust’s Clone trait, along with sharing a few similarities to the std::borrow::ToOwned trait with regards to the output type.

See Apple’s documentation for details.

Provided Methods

fn copy(&self) -> Retained<Self::Immutable>
where Self: Sized + Message + CounterpartOrSelf,

Returns a new instance that’s a copy of the receiver.

The output type is the immutable counterpart of the object, which is usually Self, but e.g. NSMutableString returns NSString.