objc2_foundation

Type Alias NSRect

pub type NSRect = CGRect;

A rectangle.

This is a convenience alias for CGRect. For ease of use, it is available on all platforms, though in practice it is only useful on macOS.

See Apple’s documentation.

Aliased Type

#[repr(C)]
pub struct NSRect {
    pub origin: CGPoint,
    pub size: CGSize,
}

Fields
origin: CGPoint

The coordinates of the rectangle’s origin.

size: CGSize

The dimensions of the rectangle.